### 2024-05-20 #6765 @editorialbot
[PRE REVIEW]: Additive Bayesian Networks
**Submitting author:** @matteodelucchi (<a href="http://orcid.org/0000-0002-9327-1496">Matteo Delucchi</a>)
**Repository:** https://github.com/furrer-lab/abn
**Branch with paper.md** (empty if default branch):
**Version:** 3.1.0
