### 2024-05-20 #6765 @crvernon
:wave: @matteodelucchi - Thanks for your submission to JOSS!  Could you provide a more detail in terms of your contributions (additions, changes) to the package since this it was published in the following:

> Kratzer, G., Lewis, F., Comin, A., Pittavino, M., & Furrer, R. (2023). Additive Bayesian Network Modeling with the R Package abn. Journal of Statistical Software, 105(8), 1–41. https://doi.org/10.18637/jss.v105.i08

I only see the commit record on GitHub going back to when you uploaded the package itself as v3.0 so I cannot run a diff from v2.* to your new version 3.
