### 2024-05-21 77d4e43
fixed badge for CRAN like checks and added JOSS status badge.
