### 2024-05-21 #6765 @matteodelucchi
> Thanks for your submission to JOSS! Could you provide a more detail in terms of your contributions (additions, changes) to the package since this it was published in the following:
>
> > Kratzer, G., Lewis, F., Comin, A., Pittavino, M., & Furrer, R. (2023). Additive Bayesian Network Modeling with the R Package abn. Journal of Statistical Software, 105(8), 1–41. https://doi.org/10.18637/jss.v105.i08
>
> I only see the commit record on GitHub going back to when you uploaded the package itself as v3.0 so I cannot run a diff from v2.* to your new version 3.

👋 @crvernon  Thank you for your comment.

Since the publication that you mentioned from Kratzer et al. 2023, we have made significant contributions to the `abn` package. The commit record is censored because, prior to version 3.0, the package was hosted on an internal GitLab Server. However, you can run a diff from the [archive on CRAN](https://cran.r-project.org/src/contrib/Archive/abn/).

In the paper by Kratzer et al. 2023, the focus was on the theoretical aspects of additive Bayesian Networks. They introduced the random-effects method under the Bayesian framework but not for the frequentist approach. In our work, we have expanded the scope to include the information-theoretic formulation.

The initial code development did not strictly adhere to best practices and was somewhat unstructured. I will try to summarise our improvements in this regard:

- **Testing**: The initial version had only a few, rather unspecific tests. We have implemented comprehensive tests, including functional tests, to ensure the package's robustness.
- **Rigorous Testing Framework**: We have implemented a rigorous testing framework that includes dependencies, allowing us to test the package on various system setups.
- **Documentation**: The documentation of functions was deprecated and partially incorrect. We have fixed this and upgraded it to follow recent standards for better clarity and maintenance. We also launched a new website for the package.
- **Method Extension**: The package has been extended with mixed-effects under the “mle” method, broadening its applicability and harmonizing its functionalities across the two concepts (`method = 'bayes'` and `method = 'mle'`).
- **Parallelization**: We implemented parallelization of `buildScoreCache(method='bayes')` and `fitAbn(method='bayes')`.
- **Code Refactoring**: We harmonized the code for the functions `buildScoreCache()` and `fitAbn()` for both methods “mle” and “bayes”.
- **Bug Fixes**: We detected various bugs along the way, such as those caused by illogical combinations of function arguments. These have now been caught and fixed.

I hope this clarifies the differences between the two publications.
