### 2024-05-22 9f913f3
cleaned up NEWS.md
