### 2024-05-23 55ba4eb
fix: fetching correct tag for previous version
