### 2024-05-23 b95a530
fix: fetching correct tag for previous version
