### 2024-05-27 #6765 @robmoss
> 👋 @robmoss - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I've had a look at the article and the [package website](https://r-bayesian-networks.org/), and I don't think I have enough familiarity/expertise with the topic.
