### 2024-05-31 #6765 @matteodelucchi
> 👋 @matteodelucchi - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@crvernon I need to decline to review because I have a conflict of interest - it's my submission :wink:
