### 2024-06-01 #6822 @editorialbot
[REVIEW]: Additive Bayesian Networks
**Submitting author:** @matteodelucchi (<a href="http://orcid.org/0000-0002-9327-1496">Matteo Delucchi</a>)
**Repository:** https://github.com/furrer-lab/abn
**Branch with paper.md** (empty if default branch):
**Version:** 3.1.3
