### 2024-06-09 furrer-lab/abn#114 @abhishektiwari
Provide guidance on prerequisite for installing abn
I faced several issues to install `abn` on both Mac 14.4.1 (M2) and Ubuntu 22.04 due to missing prerequisite packages. Please provide prerequisite guidance for users on Linux, Mac and Windows.

In case of Ubuntu 22.04, I have to install `cmake` `r-cran-rcpp` `r-cran-rcppeigen ` `r-cran-rjags` `libgsl-dev` (after several trial and error) still no luck. Basically, installation instructions don't work out of box.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6822)
