### 2024-06-09 #6822 @abhishektiwari
@matteodelucchi I am unable to progress with review as installation instruction don't work as described. Please see [this ticket](https://github.com/furrer-lab/abn/issues/114)
