### 2024-06-10 08e8651
include note on GSL and JAGS requirements for Linux and MacOS. Addressing #114
