+++ README.md
@@ -40,6 +40,19 @@ We have implemented a rigorous testing framework similar to CRAN's to validate t
 Our aim is to maximize the reliability of the `abn` package under various conditions, and we are dedicated to providing a robust and reliable package. 
 We appreciate your understanding as we work towards making `abn` available on CRAN soon.
 
+## Required System libraries
+
+The following system libraries are required to install the [abn](https://cran.r-project.org/package=abn) package:
+
+- [GSL](https://www.gnu.org/software/gsl/): The GNU Scientific Library is a numerical library for C/C++. It is required for the [abn](https://cran.r-project.org/package=abn) package to compile the C/C++ code.
+    - On Ubuntu, install the library with `sudo apt-get install libgsl-dev`.
+    - On Fedora, install the library with `sudo dnf install gsl-devel`.
+    - On macOS, install the library with `brew install gsl`.
+- [JAGS](https://mcmc-jags.sourceforge.io/): Just Another Gibbs Sampler is a program for analyzing Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. It is required for the [abn](https://cran.r-project.org/package=abn) package to run simulations.
+    - On Ubuntu, install the library with `sudo apt-get install jags`.
+    - On Fedora, install the library with `sudo dnf install jags`.
+    - On macOS, install the library with `brew install jags`.
+
 ## Additional libraries
 
 The following additional libraries are recommended to best profit from the [abn](https://cran.r-project.org/package=abn) features.
