### 2024-06-10 9bd16e1
include note on GSL and JAGS requirements for Windows. Addressing #114
