### 2024-06-10 e76296f
check if jags library is available during compilation. Addressing #114.
