### 2024-06-24 6afdbf9
badge with link to codecov report page. Addressing Setup sensible code coverage report  #109
