### 2024-06-24 26d6ccb
dev: running code coverage with all tests
