+++ .github/workflows/onlabel_CRAN_checks.yml
@@ -224,7 +224,7 @@ jobs:
           covr::codecov(
             quiet = FALSE,
             clean = FALSE,
-            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package")
+            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package"),
             token = Sys.getenv("CODECOV_TOKEN")
           )
         shell: Rscript {0}
