### 2024-06-28 furrer-lab/abn#114 @j-i-l
Hey @abhishektiwari,

were you able to finish the installation process with the provided information? Let us know if you encountered some further roadblocks, or if you think we should detail the installation process more in-depth.
