### 2024-06-30 23332ed
first implementation of proposed improved cache structure in #58. WIP, still errors.
