+++ R/build_score_cache_mle.R
@@ -27,36 +27,36 @@ forLoopContent <-
       if (verbose) {
         message("we have grouping (mixed-effects model) glmm")
       }
-    ###
-    # Build arguments for glmm.bic
-    ###
-    # current node's name
-    child <- mycache[["children"]][row.num] # child as integer
-    child.name <- colnames(mycache[["node.defn"]])[child]
-    # current node's parents names
-    parents.names <- names(which(mycache[["node.defn"]][row.num,] != 0))
-    # current node's distribution
-    child.dist <- data.dists[child] #       distribution <- data.dists[child]
-    # data.df including group.var=group.ids
-    data.df.grouping <- data.frame(data.df, grp = factor(group.ids))
-    names(data.df.grouping) <- c(names(data.df), group.var)
+      ###
+      # Build arguments for glmm.bic
+      ###
+      # current node's name
+      child <- mycache[["children"]][row.num] # child as integer
+      child.name <- colnames(mycache[["node.defn"]])[child]
+      # current node's parents names
+      parents.names <- names(which(mycache[["node.defn"]][row.num,] != 0))
+      # current node's distribution
+      child.dist <- data.dists[child] #       distribution <- data.dists[child]
+      # data.df including group.var=group.ids
+      data.df.grouping <- data.frame(data.df, grp = factor(group.ids))
+      names(data.df.grouping) <- c(names(data.df), group.var)
 
-    ###
-    # Beginn of GLMM
-    ###
-    # Build formula statement
-    if (length(parents.names) == 0){
-      # no parent only random effect
-      model <- as.formula(paste(child.name, "~ 1 + (1|", group.var, ")"))
-    } else if (length(parents.names) > 0){
-      model <- as.formula(paste(child.name, "~ (1|", group.var, ")+", paste(parents.names, collapse = "+")))
-    } else {
-      stop("Cannot build model formula. Unknown predictor names of parent nodes in 'parents.name'.")
-    }
+      ###
+      # Beginn of GLMM
+      ###
+      # Build formula statement
+      if (length(parents.names) == 0){
+        # no parent only random effect
+        model <- as.formula(paste(child.name, "~ 1 + (1|", group.var, ")"))
+      } else if (length(parents.names) > 0){
+        model <- as.formula(paste(child.name, "~ (1|", group.var, ")+", paste(parents.names, collapse = "+")))
+      } else {
+        stop("Cannot build model formula. Unknown predictor names of parent nodes in 'parents.name'.")
+      }
 
-    # Main part: Depending on child's distribution, call the appropriate modelling function
-    switch (as.character(child.dist),
-            gaussian = {
+      # Main part: Depending on child's distribution, call the appropriate modelling function
+      switch (as.character(child.dist),
+              gaussian = {
                 tryCatch({
                   fit <- lme4::lmer(model, data = data.df.grouping)
                 }, error=function(e)NULL)
@@ -75,8 +75,8 @@ forLoopContent <-
                 }
 
                 # if fit is still NULL, try fixed effect only
-            },
-            binomial = {
+              },
+              binomial = {
                 tryCatch({
                   fit <- lme4::glmer(model, data = data.df.grouping, family = "binomial")
                 }, error=function(e)NULL)
@@ -96,8 +96,8 @@ forLoopContent <-
                 }
 
                 # if fit is still NULL, do not modify model further as this would change the structure of the dag but return very low score (further down)
-            },
[…]
