### 2024-07-03 e87a154
fixing running condition
