+++ .github/workflows/onlabel_OS_checks.yml
@@ -20,8 +20,8 @@ permissions:
 
 jobs:
   R-CMD-check:
+    if: ${{ github.event.label.name == 'OS::check' }}
     name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-    if: ${{ ( github.event.label.name == needs.set_target.outputs.lint_label ) }}
     runs-on: ${{ matrix.config.os }}
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
