### 2024-07-03 a1da57d
installing dependencies for macos
