### 2024-07-03 b6678f5
try installing gsl package
