### 2024-07-04 #6822 @matteodelucchi
> 👋 @matteodelucchi, @dhvalden, and @abhishektiwari - just checking in to see how things are going with this review. Could you each provide a short update response in this thread? Thanks!

Hi @crvernon, We tried to [clarify the installation instructions](https://github.com/furrer-lab/abn/issues/114), and @abhishektiwari [mentioned](https://github.com/furrer-lab/abn/issues/114#issuecomment-2197672666) giving it another try within these days.
