### 2024-07-08 f2ff8d9
dedicated window pipeline
