+++ .github/workflows/onlabel_Windows_checks.yml
@@ -11,6 +11,7 @@ env:
   LABEL_SUCCESS: 'Windows::passed'
   LABEL_FAILURE: 'Windows::failed'
   BUILD_LOC: "./build"
+  OS: 'windows-latest'
 
 permissions:
   packages: read
@@ -21,12 +22,11 @@ permissions:
 jobs:
   R-CMD-check:
     if: ${{ github.event.label.name == 'Windows::check' }}
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-    runs-on: ${{ matrix.config.os }}
+    name: ${{ env.OS }} (${{ matrix.config.r }})
+    runs-on: ${{ env.OS }}
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-      os: windows-latest
     strategy:
       fail-fast: false
       matrix:
