+++ .github/workflows/onlabel_Macos_checks.yml
@@ -7,10 +7,11 @@ on:
     types: [ labeled ]
 
 env:
-  LABEL_CHECK: 'OS::check'
-  LABEL_SUCCESS: 'OS::passed'
-  LABEL_FAILURE: 'OS::failed'
+  LABEL_CHECK: 'macos::check'
+  LABEL_SUCCESS: 'macos::passed'
+  LABEL_FAILURE: 'macos::failed'
   BUILD_LOC: "./build"
+  OS: 'macos-latest'
 
 permissions:
   packages: read
@@ -20,9 +21,9 @@ permissions:
 
 jobs:
   R-CMD-check:
-    if: ${{ github.event.label.name == 'OS::check' }}
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-    runs-on: ${{ matrix.config.os }}
+    if: ${{ github.event.label.name == 'macos::check' }}
+    name: ${{ env.OS }} (${{ matrix.config.r }})
+    runs-on: ${{ env.OS }}
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
@@ -30,25 +31,19 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macos-latest,   r: 'release'}
-
-          - {os: windows-latest, r: 'release'}
-          # use 4.1 to check with rtools40's older compiler
-          # - {os: windows-latest, r: '4.1'}
-
-          # - {os: ubuntu-latest,  r: 'devel', http-user-agent: 'release'}
-          # - {os: ubuntu-latest,  r: 'release'}
-          # - {os: ubuntu-latest,  r: 'oldrel-1'}
-          # - {os: ubuntu-latest,  r: 'oldrel-2'}
-          # - {os: ubuntu-latest,  r: 'oldrel-3'}
-          # - {os: ubuntu-latest,  r: 'oldrel-4'}
+          - {r: 'release',}
+          - {r: '4.1',}
+          - {r: 'devel', http-user-agent: 'release'}
+          - {r: 'oldrel-1',}
+          - {r: 'oldrel-2',}
+          - {r: 'oldrel-3',}
 
     steps:
-      # - name: install dependencies macos
-      #   if: ${{ matrix.config.os == 'macos-latest' }}
-      #   run: |
-      #     brew install gsl
-      #     brew install jags
+
+      - name: install dependencies macos
+        run: |
+          brew install gsl
+          brew install jags
 
       - uses: actions/checkout@v4
 
