### 2024-07-08 5724c61
Revert "try installing gsl package"
This reverts commit b6678f53fe08036b47876b9b4dbfd3bbcd90e578.
