+++ DESCRIPTION
@@ -51,7 +51,6 @@ Imports:
     Rgraphviz,
     rjags,
     stringi
-    gsl
 Suggests: 
     bookdown,
     boot,
