+++ .github/workflows/onlabel_Windows_checks.yml
@@ -7,9 +7,9 @@ on:
     types: [ labeled ]
 
 env:
-  LABEL_CHECK: 'Windows::check'
-  LABEL_SUCCESS: 'Windows::passed'
-  LABEL_FAILURE: 'Windows::failed'
+  LABEL_CHECK: 'windows::check'
+  LABEL_SUCCESS: 'windows::passed'
+  LABEL_FAILURE: 'windows::failed'
   BUILD_LOC: "./build"
   OS: 'windows-latest'
 
@@ -21,7 +21,7 @@ permissions:
 
 jobs:
   R-CMD-check:
-    if: ${{ github.event.label.name == 'Windows::check' }}
+    if: ${{ github.event.label.name == 'windows::check' }}
     name: ${{ env.OS }} (${{ matrix.config.r }})
     runs-on: ${{ env.OS }}
     env:
