### 2024-07-08 47c81a2
installing jags
