### 2024-07-08 7bf00df
adding jags path
