+++ .github/workflows/onlabel_Windows_checks.yml
@@ -48,8 +48,9 @@ jobs:
       - name: Install JAGS
         run: |
           Import-Module bitstransfer
-          start-bitstransfer -source ${{ env.JAGS_URL }}/${{ env.JAGS_MAJORV }}.x/Windows/JAGS-${{ env.JAGS_VERSION }}.exe/download  C:\JAGS-${{ env.JAGS_VERSION }}.exe
-          & "C:\JAGS-${{ env.JAGS_VERSION }}.exe"
+          New-Item -ItemType Directory -Force -Path "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}"
+          start-bitstransfer -source ${{ env.JAGS_URL }}/${{ env.JAGS_MAJORV }}.x/Windows/JAGS-${{ env.JAGS_VERSION }}.exe/download  "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}\JAGS-${{ env.JAGS_VERSION }}.exe"
+          Sys.setenv(JAGS_HOME="C:/Program Files/JAGS/JAGS-${{ env.JAGS_VERSION }}")
           where jags
           jags
 
