### 2024-07-08 935f6b1
installing rjags providing Jags home
