### 2024-07-09 furrer-lab/abn#123 @dhvalden
Incorrect Jags command for Fedora Linux
There is no binaries for Fedora in the dnf repos. Jags for fedora has to be downloaded
