### 2024-07-09 furrer-lab/abn#124 @dhvalden
INLA Package not in CRAN
Installation stops with error: `Error: package 'INLA' is not available`

available from another repo here: https://www.r-inla.org/download-install

Need to be added to the installation.

either is included to install automatically or deleted from the dependencies, Otherwise the installation will keep stopping with error code.
