### 2024-07-09 furrer-lab/abn#125 @dhvalden
New unlisted system dependency udunits2
While installing INLA:
```
  Configuration failed because libudunits2.so was not found. Try installing:
    * deb: libudunits2-dev (Debian, Ubuntu, ...)
    * rpm: udunits2-devel (Fedora, EPEL, ...)
    * brew: udunits (OSX)
```
While is a ependency of INLA, it carries to abn since INLA is a dependency of abn.
