### 2024-07-09 5b858b3
debugging - show gls configs
