### 2024-07-09 #6822 @dhvalden
Hey I have been trying to install the package in a Fedora 40 Workstation machine. I stopped trying when I found several problems listed in the following issues:

- https://github.com/furrer-lab/abn/issues/123
- https://github.com/furrer-lab/abn/issues/124
- https://github.com/furrer-lab/abn/issues/125

I'll try again when i have some response about them (not necessary the solution, just advice of the best course to follow).
