### 2024-07-09 d458de5
create configure file with autoconf
