+++ .github/workflows/onlabel_Macos_checks.yml
@@ -60,6 +60,10 @@ jobs:
       #   run: |
       #       install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
 
+      - name: Configure ABN for installation
+        run: |
+          autoconf
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::rcmdcheck
