### 2024-07-09 0bd4348
debugging - gsl path
