### 2024-07-09 2553b5e
providing flags for clang/gcc as well
