+++ src/Makevars.in
@@ -13,7 +13,7 @@ PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_CFLAGS) -I.
 
 PKG_LIBS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_LIBS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
 
-#PKG_CFLAGS = $(SHLIB_OPENMP_CFLAGS)
+PKG_CFLAGS = $(SHLIB_OPENMP_CFLAGS) $(GSL_CFLAGS)
 
-#PKG_LIBS = $(SHLIB_OPENMP_CFLAGS)
+PKG_LIBS = $(SHLIB_OPENMP_CFLAGS)
 
