### 2024-07-09 b3230bc
debugging - check jags location
