### 2024-07-09 97033ce
try to detect the paths for jags
