+++ src/Makevars.in
@@ -3,17 +3,16 @@
 # set by configure
 
 GSL_CFLAGS = @GSL_CFLAGS@
-
 GSL_LIBS = @GSL_LIBS@
+
+JAGS_CFLAGS = @JAGS_CFLAGS@
+JAGS_LIBS = @JAGS_LIBS@
 
 
 # combine to standard arguments for R
 
-PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_CFLAGS) -I.
+PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_CFLAGS) $(JAGS_CFLAGS) -I.
 
-PKG_LIBS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_LIBS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
+PKG_LIBS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_LIBS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) $(JAGS_LIBS)
 
-PKG_CFLAGS = $(SHLIB_OPENMP_CFLAGS) $(GSL_CFLAGS)
-
-PKG_LIBS = $(SHLIB_OPENMP_CFLAGS)
-
+PKG_CFLAGS = $(SHLIB_OPENMP_CFLAGS) $(GSL_CFLAGS) $(JAGS_CFLAGS)
