+++ src/Makevars.in
@@ -12,7 +12,5 @@ JAGS_LIBS = @JAGS_LIBS@
 # combine to standard arguments for R
 
 PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_CFLAGS) $(JAGS_CFLAGS) -I.
-
-PKG_LIBS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_LIBS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) $(JAGS_LIBS)
-
 PKG_CFLAGS = $(SHLIB_OPENMP_CFLAGS) $(GSL_CFLAGS) $(JAGS_CFLAGS)
+PKG_LIBS = $(SHLIB_OPENMP_CXXFLAGS) $(GSL_LIBS) $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS) $(JAGS_LIBS)
