### 2024-07-10 ceef3ca
providing jags paths does not help
