### 2024-07-10 44bf34c
started with restructuring installation section
