### 2024-07-10 #6822 @j-i-l
> Hey I have been trying to install the package in a Fedora 40 Workstation machine. I stopped trying when I found several problems listed in the following issues:
>
>     * [Incorrect Jags command for Fedora Linux furrer-lab/abn#123](https://github.com/furrer-lab/abn/issues/123)
>
>     * [INLA Package not in CRAN furrer-lab/abn#124](https://github.com/furrer-lab/abn/issues/124)
>
>     * [New unlisted system dependency udunits2 furrer-lab/abn#125](https://github.com/furrer-lab/abn/issues/125)
>
>
> I'll try again when i have some response about them (not necessary the solution, just advice of the best course to follow).

Hi @dhvalden ,

Thanks a lot for the detailed feedback!
It seems to me that we had neglected a bit the ease of use of our package. My apologies for that!

I've opened a new pull request (https://github.com/furrer-lab/abn/pull/126) in which I'll address the issues you raised. I'd kindly invite you to try the installation process again once the pull request is merged.
