### 2024-07-10 1b84605
adding installing ... sections
