+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -7,9 +7,9 @@ on:
     types: [ labeled ]
 
 env:
-  LABEL_CHECK: 'fedora::check'
-  LABEL_SUCCESS: 'fedora::passed'
-  LABEL_FAILURE: 'fedora::failed'
+  LABEL_CHECK: 'fedoraInstall::check'
+  LABEL_SUCCESS: 'fedoraInstall::passed'
+  LABEL_FAILURE: 'fedoraInstall::failed'
   JAGS: '4.3.2'
   BUILD_LOC: "./build"
 
@@ -20,8 +20,8 @@ permissions:
   repository-projects: write
 
 jobs:
-  R-CMD-check:
-    if: ${{ github.event.label.name == 'fedora::check' }}
+  setup-and-install:
+    if: ${{ github.event.label.name == 'fedoraInstall::check' }}
     name: fedora-latest (${{ matrix.config.r }})
     runs-on: fedora-latest
     env:
@@ -31,7 +31,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {r: 'release', compiler: "clang"}
+          - {r: 'release'}
           # - {r: '4.1',}
           # - {r: 'devel', http-user-agent: 'release'}
           # - {r: 'oldrel-1',}
@@ -39,72 +39,81 @@ jobs:
           # - {r: 'oldrel-3',}
 
     steps:
-      - name: install lapack blas an atlas
+      - name: install gcc
         run: |
-          dnf -y install lapack*
-          dnf -y install blas*
-          dnf -y install atlas*
+          dnf -y install gcc
+
+       # - name: install lapack blas and atlas
+       #   run: |
+       #     dnf -y install lapack*
+       #     dnf -y install blas*
+       #     dnf -y install atlas*
+
+      - name: install GSL
+        run: |
+          dnf -y install gsl-devel
 
       - name: install JAGS
         run: |
-          wget https://sourceforge.net/projects/mcmc-jags/files/JAGS/4.x/Source/${{ env.JAGS }}.tar.gz/download /tmp/jags.tar.gz
+          wget -O /tmp/jags.tar.gz https://sourceforge.net/projects/mcmc-jags/files/JAGS/4.x/Source/JAGS-${{ env.JAGS }}.tar.gz/download
           cd /tmp
           tar -xf jags.tar.gz
-          cd /tmp/${{ env.JAGS }} 
-          if [ "${{ matrix.config.compiler}}" = "clang" ]; then RCC=/usr/bin/clang RCXX="/usr/bin/clang++ -std=gnu++17 -stdlib=libc++"; else RRCC=/usr/bin/gcc RCXX="/usr/bin/g++"; fi
-          LD="llvm-ld" CC="${RCC}" CXX="${RCXX}" ./configure --with-blas="-L/opt/R-devel/lib64/R/lib -lRblas" --with-lapack="-L/opt/R-devel/lib64/R/lib -lRlapack" --libdir=/usr/local/lib64
+          cd /tmp/JAGS-${{ env.JAGS }} 
+          ./configure --libdir=/usr/local/lib64
           make
           make install
 
-      - name: install rjags separately
+      - name: install R
         run: |
-          install.packages('rjags', configure.args = '--enable-rpath')
+          dnf -y install R
+
+      - name: install config and build dependencies
+        run: |
+          dnf -y install cmake
+          # dnf -y install autoconf
+          dnf -y install R-devtools
+
+      # - uses: actions/checkout@v4
+
+      # - name: Configure abn for installation
+      #   run: |
+      #     autoconf
+
+      - name: installing INLA
+        run: |
+            install.packages("INLA", repos = "", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
         shell: Rscript {0}
 
-      - name: install devtools
[…]
