### 2024-07-10 c9e24fa
installing missing tools
