### 2024-07-10 deb6cff
missing gfortran and deps
