### 2024-07-10 d44ab9f
more details on jags
