### 2024-07-10 af68530
fixing installation command
