+++ .github/workflows/onlabel_Macos_setup.yml
@@ -1,15 +1,15 @@
 # Workflow is triggered on label and check installation on various operating systems
 # This action is adapted from https://github.com/t4d-gmbh/stubbed_versioning
-name: On Label Macos Checks
+name: Macos Install
 
 on:
   pull_request:
     types: [ labeled ]
 
 env:
-  LABEL_CHECK: 'macos::check'
-  LABEL_SUCCESS: 'macos::passed'
-  LABEL_FAILURE: 'macos::failed'
+  LABEL_CHECK: 'macosInstall::check'
+  LABEL_SUCCESS: 'macosInstall::passed'
+  LABEL_FAILURE: 'macosInstall::failed'
   BUILD_LOC: "./build"
 
 permissions:
@@ -19,8 +19,8 @@ permissions:
   repository-projects: write
 
 jobs:
-  R-CMD-check:
-    if: ${{ github.event.label.name == 'macos::check' }}
+  setup-and_install:
+    if: ${{ github.event.label.name == 'macosInstall::check' }}
     name: macos-latest (${{ matrix.config.r }})
     runs-on: macos-latest
     env:
@@ -38,53 +38,63 @@ jobs:
           # - {r: 'oldrel-3',}
 
     steps:
-      - name: install dependencies macos
+      - name: prepare macos
         run: |
           brew install pkg-config
-          brew install gsl
-          brew install jags
           brew install automake  # needed to run autoconf
-          gsl-config --libs
-          gsl-config --cflags
 
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
-
-      - name: install rjags separately
+      - name: install GSL
         run: |
-          install.packages('rjags', configure.args = '--enable-rpath')
+          brew install gsl
+
+      - name: install JAGS
+        run: |
+          brew install jags
+
+      # installing INLA separately (?)
+
+      - name: installing dependiecies from BiocManager
+        run: |
+          if (!requireNamespace("BiocManager", quietly = TRUE))
+              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
+          BiocManager::install("Rgraphviz")
+          BiocManager::install("graph")
         shell: Rscript {0}
 
-      - name: Configure ABN for installation
+      - name: install abn from GitHub
         run: |
-          autoconf
+            devtools::install_github("furrer-lab/abn", ref="3.1.1")
+        shell: Rscript {0}
 
-      - name: Debugging
+      - name: Running testhat
         run: |
-          pkg-config --cflags --static jags
-          pkg-config --libs --static jags
+            res=devtools::test(abn)
+        shell: Rscript {0}
 
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          extra-packages: any::rcmdcheck
-          needs: check
 
-      - uses: r-lib/actions/check-r-package@v2
-        with:
-          upload-snapshots: false
-          build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
+      # - name: Configure ABN for installation
+      #   run: |
+      #     autoconf
 
-  report_macos_checks:
+      # - name: Debugging
+      #   run: |
+      #     pkg-config --cflags --static jags
+      #     pkg-config --libs --static jags
+
[…]
