+++ .github/workflows/onlabel_Windows_setup.yml
@@ -7,9 +7,9 @@ on:
     types: [ labeled ]
 
 env:
-  LABEL_CHECK: 'windows::check'
-  LABEL_SUCCESS: 'windows::passed'
-  LABEL_FAILURE: 'windows::failed'
+  LABEL_CHECK: 'windowsInstall::check'
+  LABEL_SUCCESS: 'windowsInstall::passed'
+  LABEL_FAILURE: 'windowsInstall::failed'
   BUILD_LOC: "./build"
   JAGS_MAJORV: '4'
   JAGS_VERSION: '4.3.1'
@@ -22,8 +22,8 @@ permissions:
   repository-projects: write
 
 jobs:
-  R-CMD-check:
-    if: ${{ github.event.label.name == 'windows::check' }}
+  setup-and-install:
+    if: ${{ github.event.label.name == 'windowsInstall::check' }}
     name: windows-latest (${{ matrix.config.r }})
     runs-on: windows-latest
     env:
@@ -42,50 +42,46 @@ jobs:
 
     steps:
 
-      - name: Install GSL binaries with cygwin
-        uses: cygwin/cygwin-install-action@master
+      # - name: Install GSL binaries with cygwin
+      #   uses: cygwin/cygwin-install-action@master
 
-      - name: Install JAGS
+      - name: install GSL
+        run: |
+          Import-Module bitstransfer
+          New-Item -ItemType Directory -Force -Path "C:\Program Files\cygwin"
+          start-bitstransfer -source https://cygwin.com/setup-x86_64.exe "C:\Program Files\cygwin\setup-x86_64.exe"
+          Start-Process -Wait -FilePath "C:\Program Files\cygwin\setup-x86_64.exe" -ArgumentList "/S" -PassThru
+
+      - name: install JAGS
         run: |
           Import-Module bitstransfer
           New-Item -ItemType Directory -Force -Path "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}"
           start-bitstransfer -source ${{ env.JAGS_URL }}/${{ env.JAGS_MAJORV }}.x/Windows/JAGS-${{ env.JAGS_VERSION }}.exe/download  "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}\JAGS-${{ env.JAGS_VERSION }}.exe"
           Start-Process -Wait -FilePath "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}\JAGS-${{ env.JAGS_VERSION }}.exe" -ArgumentList "/S" -PassThru
 
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
-
-      - name: install JAGS
+      - name: install rjags separately
         run: |
             Sys.setenv(JAGS_HOME="C:/Program Files/JAGS/JAGS-${{ env.JAGS_VERSION }}")
             install.packages("rjags")
         shell: Rscript {0}
 
-      - name: install JAGS
+      # - name: install INLA
+      #   run: |
+      #       install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+
+      - name: install abn from GitHub
         run: |
-            install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+            devtools::install_github("furrer-lab/abn", ref="3.1.1")
+        shell: Rscript {0}
 
[…]
