### 2024-07-10 7cde62f
separately install rjags
