+++ .github/workflows/onlabel_Macos_setup.yml
@@ -68,7 +68,7 @@ jobs:
 
       - name: Running testhat
         run: |
-            res=devtools::test(abn)
+            testthat::test_package('abn')
         shell: Rscript {0}
 
 
