+++ .github/workflows/onlabel_Windows_setup.yml
@@ -88,7 +88,7 @@ jobs:
 
       - name: Running testhat
         run: |
-            res=devtools::test(abn)
+            testthat::test_package('abn')
         shell: Rscript {0}
   report_Windows_setup:
     if: ${{ (success() || failure()) }}
