### 2024-07-10 bd57b51
installing INLA separately
