+++ .github/workflows/onlabel_Windows_setup.yml
@@ -69,9 +69,10 @@ jobs:
             install.packages("rjags", repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
 
-      # - name: install INLA
-      #   run: |
-      #       install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+      - name: install INLA
+        run: |
+            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+        shell: Rscript {0}
 
       - name: installing dependiecies from BiocManager
         run: |
