+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -87,7 +87,10 @@ jobs:
       #   run: |
       #     autoconf
 
-      # installing INLA separately (?)
+      - name: install INLA
+        run: |
+            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+        shell: Rscript {0}
 
       - name: installing dependiecies from BiocManager
         run: |
@@ -113,7 +116,7 @@ jobs:
 
       - name: Running testhat
         run: |
-            res=devtools::test(abn)
+            testthat::test_package('abn')
         shell: Rscript {0}
 
   report_Fedora_setup:
