+++ .github/workflows/onlabel_Macos_setup.yml
@@ -51,7 +51,10 @@ jobs:
         run: |
           brew install jags
 
-      # installing INLA separately (?)
+      - name: install INLA
+        run: |
+            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+        shell: Rscript {0}
 
       - name: installing dependiecies from BiocManager
         run: |
