+++ .github/workflows/onlabel_Macos_setup.yml
@@ -40,6 +40,7 @@ jobs:
     steps:
       - name: prepare macos
         run: |
+          brew install R
           brew install pkg-config
           brew install automake  # needed to run autoconf
 
