### 2024-07-10 60ace2e
test without cmake
