### 2024-07-10 2d0bd77
getting the tests and run testthat
