+++ .github/workflows/onlabel_Windows_setup.yml
@@ -44,10 +44,22 @@ jobs:
 
       # - name: Install GSL binaries with cygwin
       #   uses: cygwin/cygwin-install-action@master
-      - name: prepare windows
+      - name: prepare windows 1/2
         run: |
             install.packages("devtools", repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
+
+      # - name: prepare windows 2/2
+      #   run: |
+      #       $progressPreference = 'silentlyContinue'
+      #       Write-Information "Downloading WinGet and its dependencies..."
+      #       Invoke-WebRequest -Uri https://aka.ms/getwinget -OutFile Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle
+      #       Invoke-WebRequest -Uri https://aka.ms/Microsoft.VCLibs.x64.14.00.Desktop.appx -OutFile Microsoft.VCLibs.x64.14.00.Desktop.appx
+      #       Invoke-WebRequest -Uri https://github.com/microsoft/microsoft-ui-xaml/releases/download/v2.8.6/Microsoft.UI.Xaml.2.8.x64.appx -OutFile Microsoft.UI.Xaml.2.8.x64.appx
+      #       Add-AppxPackage Microsoft.VCLibs.x64.14.00.Desktop.appx
+      #       Add-AppxPackage Microsoft.UI.Xaml.2.8.x64.appx
+      #       Add-AppxPackage Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle
+      #       winget install --id Git.Git -e --source winget
 
       - name: install GSL
         run: |
@@ -87,10 +99,14 @@ jobs:
             devtools::install_github("furrer-lab/abn", ref="3.1.1")
         shell: Rscript {0}
 
-      - name: Running testhat
+      - name: fetch the repository from GitHub
+        uses: actions/checkout@v4
+
+      - name: running testhat
         run: |
-            testthat::test_package('abn')
+            devtools::test('./')
         shell: Rscript {0}
+
   report_Windows_setup:
     if: ${{ (success() || failure()) }}
     needs:
