+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -114,9 +114,12 @@ jobs:
       #     devtools::install('./')
       #   shell: Rscript {0}
 
-      - name: Running testhat
+      - name: fetch the repository from GitHub
+        uses: actions/checkout@v4
+
+      - name: running testhat
         run: |
-            testthat::test_package('abn')
+            devtools::test('./')
         shell: Rscript {0}
 
   report_Fedora_setup:
