+++ .github/workflows/onlabel_Windows_setup.yml
@@ -86,7 +86,7 @@ jobs:
             install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
         shell: Rscript {0}
 
-      - name: installing dependiecies from BiocManager
+      - name: installing dependencies from BiocManager
         run: |
           if (!requireNamespace("BiocManager", quietly = TRUE))
               install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
@@ -101,6 +101,11 @@ jobs:
 
       - name: fetch the repository from GitHub
         uses: actions/checkout@v4
+
+      - name: deactivate the renv
+        run: |
+            renv::deactivate()
+        shell: Rscript {0}
 
       - name: running testhat
         run: |
