### 2024-07-10 6d6dc4a
fixing path to jags.pc
