+++ .github/workflows/onlabel_Macos_setup.yml
@@ -67,7 +67,7 @@ jobs:
 
       - name: install abn from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="3.1.1")
+            devtools::install_github("furrer-lab/abn", ref="3.1.1", dependencies=c("Depends", "Imports", "Suggests", "LinkingTo"))
         shell: Rscript {0}
 
       - name: fetch the repository from GitHub
