### 2024-07-10 4ed0aa1
also install rjags explicitly
