+++ .github/workflows/onlabel_Macos_setup.yml
@@ -52,6 +52,11 @@ jobs:
         run: |
           brew install jags
 
+      - name: install rjags
+        run: |
+            install.packages("rjags", type="source", repos=c(CRAN="https://cran.r-project.org"))
+            library("rjags")
+
       - name: install INLA
         run: |
             install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
