### 2024-07-10 33f8949
install pkg-config for rjags
