### 2024-07-10 c09f9e3
working config on fedora server
