### 2024-07-10 be0c4ef
try with the pkg-config path
