+++ .github/workflows/onlabel_Windows_setup.yml
@@ -79,6 +79,7 @@ jobs:
         run: |
             Sys.setenv(JAGS_HOME="C:/Program Files/JAGS/JAGS-${{ env.JAGS_VERSION }}")
             install.packages("rjags", repos=c(CRAN="https://cran.r-project.org"))
+            library("rjags")
         shell: Rscript {0}
 
       - name: install INLA
