### 2024-07-10 f724d4b
provide path to jags library
