### 2024-07-10 4437722
try without pkg_config_path
