### 2024-07-10 219e0e1
details about fedora and jags
