+++ .github/workflows/onlabel_Windows_setup.yml
@@ -97,7 +97,7 @@ jobs:
 
       - name: install abn from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="3.1.1", dependencies=c("Depends", "Imports", "LinkingTo"))
+            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"))
         shell: Rscript {0}
 
       - name: fetch the repository from GitHub
