+++ .github/workflows/onlabel_Macos_setup.yml
@@ -74,7 +74,7 @@ jobs:
 
       - name: install abn from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="3.1.1", type="source", dependencies=c("Depends", "Imports", "LinkingTo"))
+            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", type="source", dependencies=c("Depends", "Imports", "LinkingTo"))
         shell: Rscript {0}
 
       - name: fetch the repository from GitHub
