### 2024-07-11 f87bcc7
run autoconf before installing from source
