+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -106,10 +106,10 @@ jobs:
           BiocManager::install("graph")
         shell: Rscript {0}
 
-      - name: install abn from GitHub
-        run: |
-            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"))
-        shell: Rscript {0}
+      # - name: install abn from GitHub
+      #   run: |
+      #       devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"))
+      #   shell: Rscript {0}
 
       # - name: Install package dependencies
       #   run: |
@@ -127,6 +127,10 @@ jobs:
         run: |
             renv::deactivate()
         shell: Rscript {0}
+
+      - name: Configure ABN for installation
+        run: |
+          autoconf
 
       - name: running testhat
         run: |
