### 2024-07-11 7971f17
started with macos prereqs
