### 2024-07-11 2ad346a
separate location for build and test
