### 2024-07-11 fae06ae
Ubuntu installation workflow
