+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -1,0 +1,180 @@
+# Workflow is triggered on label and check installation on various operating systems
+# This action is adapted from https://github.com/t4d-gmbh/stubbed_versioning
+name: Ubuntu Install
+
+on:
+  pull_request:
+    types: [ labeled ]
+
+env:
+  LABEL_CHECK: 'ubuntuInstall::check'
+  LABEL_SUCCESS: 'ubuntuInstall::passed'
+  LABEL_FAILURE: 'ubuntuInstall::failed'
+  JAGS: '4.3.2'
+  BUILD_LOC: "./build"
+
+permissions:
+  packages: read
+  contents: write
+  pull-requests: write
+  repository-projects: write
+
+jobs:
+  setup-and-install:
+    runs-on: ubuntu-latest
+    if: ${{ github.event.label.name == 'ubuntuInstall::check' }}
+    name: ubuntu-latest (${{ matrix.config.r }})
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_KEEP_PKG_SOURCE: yes
+    strategy:
+      fail-fast: false
+      matrix:
+        config:
+          - {r: 'release'}
+          # - {r: '4.1',}
+          # - {r: 'devel', http-user-agent: 'release'}
+          # - {r: 'oldrel-1',}
+          # - {r: 'oldrel-2',}
+          # - {r: 'oldrel-3',}
+
+    steps:
+      # - name: prepare fedora
+      #   run: |
+      #     apt-get -y install wget
+      #     apt-get -y install gcc
+      #     apt-get -y install gcc-c++
+      #     apt-get -y install gfortran
+      #     apt-get -y install lapack*  # is needed on fedora server
+      #     apt-get -y install blas*
+      #     apt-get -y install atlas*
+
+      - name: install R
+        run: |
+          apt-get -y install R
+
+      - name: install config and build dependencies
+        run: |
+          apt-get -y install pkg-config
+          apt-get -y install cmake
+          apt-get -y install R-devtools
+          # dnf -y install autoconf
+
+      - name: install GSL
+        run: |
+          apt-get -y install libgsl-dev
+
+      - name: install JAGS
+        run: |
+          apt-get -y install jags
+
+      - name: install rjags
+        run: |
+            install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
+            library("rjags")
+        shell: Rscript {0}
+
+      # - uses: actions/checkout@v4
+
+      # - name: Configure abn for installation
+      #   run: |
+      #     autoconf
+
+      - name: install INLA
+        run: |
+            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
+        shell: Rscript {0}
+
+      - name: installing dependencies from BiocManager
+        run: |
+          if (!requireNamespace("BiocManager", quietly = TRUE))
+              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
+          BiocManager::install("Rgraphviz")
+          BiocManager::install("graph")
+        shell: Rscript {0}
+
+      # - name: install abn from GitHub
+      #   run: |
+      #       devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"))
+      #   shell: Rscript {0}
+
+      # - name: Install package dependencies
+      #   run: |
[…]
