### 2024-07-11 995d8ad
debugging folder permisssions
