### 2024-07-11 94c21bd
allow adm group to write to /usr/local/lib
