+++ .github/workflows/onlabel_Macos_setup.yml
@@ -77,40 +77,43 @@ jobs:
             devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", type="source", dependencies=c("Depends", "Imports", "LinkingTo"))
         shell: Rscript {0}
 
+
       - name: fetch the repository from GitHub
         uses: actions/checkout@v4
-
-      # # Install from source
-      # - name: Configure ABN for installation
-      #   run: |
-      #     autoconf
 
       - name: deactivate the renv
         run: |
             renv::deactivate()
         shell: Rscript {0}
 
+      - name: configure ABN for installation
+        run: |
+          autoconf
+
       - name: running testhat
         run: |
             devtools::test('./')
         shell: Rscript {0}
 
+      - name: build, install and test
+        run: |
+          pkg.name <- "abn"
+          pkg.root <- normalizePath(file.path("./"), mustWork = TRUE)
+          tmp.lib <- tempfile()
+          tmp.out <- tempfile()
+          dir.create(tmp.lib)
+          dir.create(tmp.out)
+          owd <- setwd(tmp.out)
+          devtools::install_deps("./", upgrade="never", repos=c(CRAN="https://cran.r-project.org"))
+          tools::Rcmd(c("build", pkg.root))
+          list.files(tmp.out)
+          tools::Rcmd(c("INSTALL", "-l", tmp.lib, "--install-tests", "*.tar.gz"))
+          tools::testInstalledPackage(pkg.name, lib.loc = tmp.lib, types = "tests")
+          list.files(tmp.out, recursive = TRUE)
+          unlink(tmp.out)
+          unlink(tmp.lib)
+        shell: Rscript {0}
 
-
-      # - name: Debugging
-      #   run: |
-      #     pkg-config --cflags --static jags
-      #     pkg-config --libs --static jags
-
-      # - uses: r-lib/actions/setup-r-dependencies@v2
-      #   with:
-      #     extra-packages: any::rcmdcheck
-      #     needs: check
-
-      # - uses: r-lib/actions/check-r-package@v2
-      #   with:
-      #     upload-snapshots: false
-      #     build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
 
   report_Macos_setup:
     if: ${{ (success() || failure()) }}
