### 2024-07-11 a4d3b87
adding dependencies for testing
