+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -128,6 +128,8 @@ jobs:
             renv::deactivate()
         shell: Rscript {0}
 
+      - uses: r-lib/actions/setup-pandoc@v2
+
       - name: configure ABN for installation
         run: |
           autoconf
