+++ .github/workflows/onlabel_Macos_setup.yml
@@ -86,6 +86,8 @@ jobs:
             renv::deactivate()
         shell: Rscript {0}
 
+      - uses: r-lib/actions/setup-pandoc@v2
+
       - name: configure ABN for installation
         run: |
           autoconf
