+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -108,6 +108,8 @@ jobs:
             renv::deactivate()
         shell: Rscript {0}
 
+      - uses: r-lib/actions/setup-pandoc@v2
+
       - name: configure ABN for installation
         run: |
           autoconf
