+++ .github/workflows/onlabel_Macos_setup.yml
@@ -59,6 +59,10 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
+      - name: install INLA dependencies
+        run: |
+          brew install udunits 
+
       - name: install INLA
         run: |
             install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
