### 2024-07-12 90b52cc
adiding curl dev for inla dep.
