+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -50,6 +50,10 @@ jobs:
       #     sudo apt-get -y install blas*
       #     sudo apt-get -y install atlas*
 
+      - name: prepare ubuntu
+        run: | 
+          sudo apt -y install libcurl4-openssl-dev
+
       - name: install R
         run: |
           sudo apt-get -y install r-base
