### 2024-07-12 30cfd5e
first make dir writeable
