### 2024-07-12 100272c
installing gdal for INLA
