### 2024-07-12 d04f4db
jpeg headers for INLA dependency
