### 2024-07-12 e90c50f
adding gdal headers for INLA
