+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -99,6 +99,7 @@ jobs:
           dnf -y install libcurl-devel  # rCurls package (for INLA)
           dnf -y install openssl-devel  # s2 package
           dnf -y install libjpeg-devel  # jpeg package
+          dnf -y install gdal-devel     # terra package
 
       - name: install INLA
         run: |
