### 2024-07-12 da6debb
try install avoiding useing local .o and .so
