+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -116,7 +116,7 @@ jobs:
 
       - name: install from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"))
+            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
         shell: Rscript {0}
 
       # - name: Install package dependencies
