+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -112,7 +112,7 @@ jobs:
 
       - name: install abn from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", type="source", dependencies=c("Depends", "Imports", "LinkingTo"))
+            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
         shell: Rscript {0}
 
 
