### 2024-07-12 3f3944c
testing the install via CRAN
