+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -109,6 +109,11 @@ jobs:
           BiocManager::install("graph")
         shell: Rscript {0}
 
+      - name: install from CRAN
+        run: |
+          install.packages('abn')
+          library('abn')
+        shell: Rscript {0}
 
       - name: install abn from GitHub
         run: |
