+++ .github/workflows/onlabel_Macos_setup.yml
@@ -64,6 +64,7 @@ jobs:
           brew install udunits 
           brew install openssl@1.1  # s2 package
           brew install gdal  # s2 package for fmesher
+          brew install proj  # sf package
 
       - name: install INLA
         run: |
