### 2024-07-12 5268ae3
specify mirror for CRAN install
