+++ .github/workflows/onlabel_Windows_setup.yml
@@ -97,7 +97,7 @@ jobs:
 
       - name: install from CRAN
         run: |
-          install.packages('abn')
+          install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
           library('abn'
         shell: Rscript {0}
 
