+++ .github/workflows/onlabel_Windows_setup.yml
@@ -98,7 +98,7 @@ jobs:
       - name: install from CRAN
         run: |
           install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
-          library('abn'
+          library('abn')
         shell: Rscript {0}
 
       - name: install abn from GitHub
