+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -123,7 +123,8 @@ jobs:
 
       - name: install from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          library('abn')
         shell: Rscript {0}
 
       # - name: Install package dependencies
