+++ .github/workflows/onlabel_Windows_setup.yml
@@ -103,7 +103,8 @@ jobs:
 
       - name: install from GitHub
         run: |
-            devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          library('abn')
         shell: Rscript {0}
 
       - name: fetch the repository from GitHub
