### 2024-07-12 b733774
cleaning up windows workflow
