+++ .github/workflows/onlabel_Windows_setup.yml
@@ -42,24 +42,10 @@ jobs:
 
     steps:
 
-      # - name: Install GSL binaries with cygwin
-      #   uses: cygwin/cygwin-install-action@master
-      - name: prepare windows 1/2
+      - name: prepare windows
         run: |
             install.packages("devtools", repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
-
-      # - name: prepare windows 2/2
-      #   run: |
-      #       $progressPreference = 'silentlyContinue'
-      #       Write-Information "Downloading WinGet and its dependencies..."
-      #       Invoke-WebRequest -Uri https://aka.ms/getwinget -OutFile Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle
-      #       Invoke-WebRequest -Uri https://aka.ms/Microsoft.VCLibs.x64.14.00.Desktop.appx -OutFile Microsoft.VCLibs.x64.14.00.Desktop.appx
-      #       Invoke-WebRequest -Uri https://github.com/microsoft/microsoft-ui-xaml/releases/download/v2.8.6/Microsoft.UI.Xaml.2.8.x64.appx -OutFile Microsoft.UI.Xaml.2.8.x64.appx
-      #       Add-AppxPackage Microsoft.VCLibs.x64.14.00.Desktop.appx
-      #       Add-AppxPackage Microsoft.UI.Xaml.2.8.x64.appx
-      #       Add-AppxPackage Microsoft.DesktopAppInstaller_8wekyb3d8bbwe.msixbundle
-      #       winget install --id Git.Git -e --source winget
 
       - name: install GSL
         run: |
