### 2024-07-12 17d5860
cleaning up macos workflow
