+++ .github/workflows/onlabel_Macos_setup.yml
@@ -54,7 +54,6 @@ jobs:
 
       - name: install rjags
         run: |
-            #Sys.setenv(JAGS_HOME="/opt/homeberw/Cellar/jags/4.3.2")
             install.packages("rjags", type="source", repos=c(CRAN="https://cran.r-project.org"))
             library("rjags")
         shell: Rscript {0}
@@ -91,7 +90,6 @@ jobs:
           library('abn')
         shell: Rscript {0}
 
-
       - name: fetch the repository from GitHub
         uses: actions/checkout@v4
 
@@ -110,26 +108,6 @@ jobs:
         run: |
             devtools::test('./')
         shell: Rscript {0}
-
-      - name: build, install and test
-        run: |
-          pkg.name <- "abn"
-          pkg.root <- normalizePath(file.path("./"), mustWork = TRUE)
-          tmp.lib <- tempfile()
-          tmp.out <- tempfile()
-          dir.create(tmp.lib)
-          dir.create(tmp.out)
-          owd <- setwd(tmp.out)
-          devtools::install_deps(pkg.root, upgrade="never", repos=c(CRAN="https://cran.r-project.org"))
-          tools::Rcmd(c("build", pkg.root))
-          list.files(tmp.out)
-          tools::Rcmd(c("INSTALL", "-l", tmp.lib, "--install-tests", "*.tar.gz"))
-          tools::testInstalledPackage(pkg.name, lib.loc = tmp.lib, types = "tests")
-          list.files(tmp.out, recursive = TRUE)
-          unlink(tmp.out)
-          unlink(tmp.lib)
-        shell: Rscript {0}
-
 
   report_Macos_setup:
     if: ${{ (success() || failure()) }}
