### 2024-07-12 b301aeb
cleaning up fedora workflow
