+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -60,7 +60,6 @@ jobs:
           dnf -y install pkg-config
           dnf -y install cmake
           dnf -y install R-devtools
-          # dnf -y install autoconf
 
       - name: install GSL
         run: |
@@ -86,12 +85,6 @@ jobs:
             install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
             library("rjags")
         shell: Rscript {0}
-
-      # - uses: actions/checkout@v4
-
-      # - name: Configure abn for installation
-      #   run: |
-      #     autoconf
 
       - name: install INLA dependencies
         run: |
@@ -127,15 +120,6 @@ jobs:
           library('abn')
         shell: Rscript {0}
 
-      # - name: Install package dependencies
-      #   run: |
-      #       devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never')
-
-      # - name: installing abn from source
-      #   run: |
-      #     devtools::install('./')
-      #   shell: Rscript {0}
-
       - name: fetch the repository from GitHub
         uses: actions/checkout@v4
 
@@ -146,38 +130,10 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - name: configure ABN for installation
-        run: |
-          autoconf
-
       - name: running testhat
         run: |
             devtools::test('./')
         shell: Rscript {0}
-
-      - name: build, install and test
-        run: |
-          pkg.name <- "abn"
-          pkg.root <- normalizePath(file.path("./"), mustWork = TRUE)
-          tmp.lib <- tempfile()
-          tmp.out <- tempfile()
-          dir.create(tmp.lib)
-          dir.create(tmp.out)
-          owd <- setwd(tmp.out)
-          devtools::install_deps(pkg.root, upgrade="never", repos=c(CRAN="https://cran.r-project.org"))
-          tools::Rcmd(c("build", pkg.root))
-          list.files(tmp.out)
-          tools::Rcmd(c("INSTALL", "-l", tmp.lib, "--install-tests", "*.tar.gz"))
-          tools::testInstalledPackage(pkg.name, lib.loc = tmp.lib, types = "tests")
-          list.files(tmp.out, recursive = TRUE)
-          unlink(tmp.out)
-          unlink(tmp.lib)
-        shell: Rscript {0}
-
-      # - name: running testhat
-      #   run: |
-      #       devtools::test('./')
-      #   shell: Rscript {0}
 
   report_Fedora_setup:
     if: ${{ (success() || failure()) }}
