### 2024-07-12 724061b
cleaning up ubuntu workflow
