+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -40,17 +40,6 @@ jobs:
 
     steps:
 
-      # - name: prepare fedora
-      #   run: |
-      #     sudo apt-get -y install wget
-      #     sudo apt-get -y install gcc
-      #     sudo apt-get -y install gcc-c++
-      #     sudo apt-get -y install gfortran
-      #     sudo apt-get -y install lapack*  # is needed on fedora server
-      #     sudo apt-get -y install blas*
-      #     sudo apt-get -y install atlas*
-
-
       - name: install R
         run: |
           sudo apt-get -y install r-base
@@ -60,7 +49,6 @@ jobs:
           sudo apt-get -y install pkg-config
           sudo apt-get -y install cmake
           sudo apt-get -y install r-cran-devtools
-          # sudo apt-get -y install autoconf
 
       - name: fixing folder perms for local lib
         run: |
@@ -87,7 +75,6 @@ jobs:
             install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
             library("rjags")
         shell: Rscript {0}
-
 
       - name: install INLA dependencies
         run: |
@@ -122,7 +109,6 @@ jobs:
           library('abn')
         shell: Rscript {0}
 
-
       - name: fetch the repository from GitHub
         uses: actions/checkout@v4
 
@@ -133,38 +119,10 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - name: configure ABN for installation
-        run: |
-          autoconf
-
       - name: running testhat
         run: |
             devtools::test('./')
         shell: Rscript {0}
-
-      - name: build, install and test
-        run: |
-          pkg.name <- "abn"
-          pkg.root <- normalizePath(file.path("./"), mustWork = TRUE)
-          tmp.lib <- tempfile()
-          tmp.out <- tempfile()
-          dir.create(tmp.lib)
-          dir.create(tmp.out)
-          owd <- setwd(tmp.out)
-          devtools::install_deps(pkg.root, upgrade="never", repos=c(CRAN="https://cran.r-project.org"))
-          tools::Rcmd(c("build", pkg.root))
-          list.files(tmp.out)
-          tools::Rcmd(c("INSTALL", "-l", tmp.lib, "--install-tests", "*.tar.gz"))
-          tools::testInstalledPackage(pkg.name, lib.loc = tmp.lib, types = "tests")
-          list.files(tmp.out, recursive = TRUE)
-          unlink(tmp.out)
-          unlink(tmp.lib)
-        shell: Rscript {0}
-
-      # - name: running testhat
-      #   run: |
-      #       devtools::test('./')
-      #   shell: Rscript {0}
 
   report_Ubuntu_setup:
     if: ${{ (success() || failure()) }}
