### 2024-07-12 622879d
installing geos>terra>INLA
