+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -80,8 +80,9 @@ jobs:
         run: |
           sudo apt-get install -y --no-install-recommends libudunits2-dev
           sudo apt -y install libcurl4-openssl-dev  # rCurl for (for INLA)
-          sudo apt-get -y install libssl-dev  # s2 package
+          sudo apt-get -y install libssl-dev   # s2 package
           sudo apt-get -y install libproj-dev  # sf package
+          sudo apt-get -y install libgeos-dev  # terra package
           sudo apt-get -y install libgdal-dev  # terra package
 
       - name: install INLA
