+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -120,9 +120,11 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - name: running testhat
+      - name: install from source
         run: |
-            devtools::test('./')
+            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never')
+            devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz')
+            install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
         shell: Rscript {0}
 
   report_Ubuntu_setup:
