+++ .github/workflows/onlabel_Macos_setup.yml
@@ -101,13 +101,11 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - name: configure ABN for installation
+      - name: install from source
         run: |
-          autoconf
-
-      - name: running testhat
-        run: |
-            devtools::test('./')
+            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never')
+            devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz')
+            install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
         shell: Rscript {0}
 
   report_Macos_setup:
