### 2024-07-13 43a8519
adding pandoc to build from source
