### 2024-07-13 e089bd3
fixing devtools info for mac +
