+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -120,7 +120,7 @@ jobs:
 
       - name: install from source
         run: |
-            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never')
+            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never', repos=c(CRAN="https://cran.r-project.org"))
             devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz', vignettes = FALSE)
             install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
         shell: Rscript {0}
