+++ .github/workflows/onlabel_Windows_setup.yml
@@ -24,22 +24,11 @@ permissions:
 jobs:
   setup-and-install:
     if: ${{ github.event.label.name == 'windowsInstall::check' }}
-    name: windows-latest (${{ matrix.config.r }})
+    name: windows-latest R release
     runs-on: windows-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          - {r: 'release',}
-          # - {r: '4.1',}
-          - {r: 'devel', http-user-agent: 'release'}
-          - {r: 'oldrel-1',}
-          - {r: 'oldrel-2',}
-          # - {r: 'oldrel-3',}
-
     steps:
 
       - name: prepare windows
