+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -23,21 +23,10 @@ jobs:
   setup-and-install:
     runs-on: ubuntu-latest
     if: ${{ github.event.label.name == 'ubuntuInstall::check' }}
-    name: ubuntu-latest (${{ matrix.config.r }})
+    name: ubuntu-latest R release
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          - {r: 'release'}
-          # - {r: '4.1',}
-          # - {r: 'devel', http-user-agent: 'release'}
-          # - {r: 'oldrel-1',}
-          # - {r: 'oldrel-2',}
-          # - {r: 'oldrel-3',}
-
     steps:
 
       - name: install R
