+++ .github/workflows/onlabel_Macos_setup.yml
@@ -21,22 +21,11 @@ permissions:
 jobs:
   setup-and-install:
     if: ${{ github.event.label.name == 'macosInstall::check' }}
-    name: macos-latest (${{ matrix.config.r }})
+    name: macos-latest R release
     runs-on: macos-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          - {r: 'release',}
-          # - {r: '4.1',}
-          # - {r: 'devel', http-user-agent: 'release'}
-          # - {r: 'oldrel-1',}
-          # - {r: 'oldrel-2',}
-          # - {r: 'oldrel-3',}
-
     steps:
       - name: prepare macos
         run: |
