### 2024-07-13 a7f44bb
removing virtual env setup for now (is untested)
