### 2024-07-13 03336c8
adding install status badges
