### 2024-07-13 furrer-lab/abn#124 @abhishektiwari
+1, unable to complete INLA installation on Mac and Ubuntu. No progress on installing `abn` so far.
