### 2024-07-13 furrer-lab/abn#114 @abhishektiwari
Unable to complete INLA installation on Mac and Ubuntu. No progress on installing `abn` so far. I am ok to not re-open this ticket in favour of other open tickets.
