### 2024-07-13 #6822 @abhishektiwari
@j-i-l I just tried to install `abn` on Mac and Ubuntu and installation failed.

First, `Error: package 'INLA' is not available.` then I followed 'INLA' installation instruction but now I have a new error `Error installing package 'nloptr'`.

@crvernon Unable to make progress unless installation instructions work. @j-i-l continue to update instructions but there are few additional hurdles as highlighted by @dhvalden.
