### 2024-07-13 furrer-lab/abn#129 @abhishektiwari
[JOSS Review] Small typo in paper 
`Statement of need` not `Statment of need`

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6822)
