+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -12,6 +12,7 @@ env:
   LABEL_FAILURE: 'ubuntuInstall::failed'
   JAGS: '4.3.2'
   BUILD_LOC: "./build"
+  BRANCH: ${{ github.ref_name }}
 
 permissions:
   packages: read
@@ -95,7 +96,7 @@ jobs:
 
       - name: install from GitHub
         run: |
-          devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
           library('abn')
         shell: Rscript {0}
 
