+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -12,6 +12,7 @@ env:
   LABEL_FAILURE: 'fedoraInstall::failed'
   JAGS: '4.3.2'
   BUILD_LOC: "./build"
+  BRANCH: ${{ github.ref_name }}
 
 permissions:
   packages: read
@@ -106,7 +107,7 @@ jobs:
 
       - name: install from GitHub
         run: |
-          devtools::install_github("furrer-lab/abn", ref="detailed-installation-instructions", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
           library('abn')
         shell: Rscript {0}
 
