+++ README.md
@@ -138,254 +138,237 @@ Some are C/C++ libraries and some are R packages that are not hosted on CRAN.
 
 Below is a list of these dependencies with details on how to install and configure them:
 
-- <details>
-    <summary><b>GSL</b></summary>
+<details>
+<summary><b>GSL</b></summary>
 
-    [GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
-    It is required to compile `abn`'s C/C++ code.
-  
-    ### Installation
+[GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
+It is required to compile `abn`'s C/C++ code.
 
-    - <details>
-        <summary>Ubuntu</summary>
-
-        `GSL` is available throught the Advanced Package Tool, `APT`, simply type:
-
-        ```bash
-        sudo apt-get install libgsl-dev
-        ```
-
-      </details>
-
-    - <details>
-        <summary>Fedora</summary>
-
-        `GSL` is available throught Fedoras standart package manager, `DNF`, simply type:
-
-        ```bash
-        sudo dnf -y install gsl-devel
-        ```
-
-      </details>
-
-    - <details>
-        <summary>MacOS</summary>
-
-        With homebrew you can install the `GSL` binaries directly:
-
-        ```
-        brew install gsl
-        ```
-      </details>
-
-    - <details>
-        <summary>Windows</summary>
-
-        In Windows `GSL` is available a.o. through [cygwin](https://cygwin.com/index.html), which has a straight forward installaiton process.
-        Eigher head over to the website, download and install the setup-x86_64.exe file, or use PowerShell:
-        
-        ```
-        Import-Module bitstransfer
-        New-Item -ItemType Directory -Force -Path "C:\Program Files\cygwin"
-        start-bitstransfer -source https://cygwin.com/setup-x86_64.exe "C:\Program Files\cygwin\setup-x86_64.exe"
-        Start-Process -Wait -FilePath "C:\Program Files\cygwin\setup-x86_64.exe" -ArgumentList "/S" -PassThru
-        ```
-
-      </details>
-  
-  </details>
+#### Installation
 
 - <details>
-    <summary><b>JAGS & rjags</b></summary>
+  <summary><b>Ubuntu</b></summary>
   
-    [JAGS](https://mcmc-jags.sourceforge.io/), _Just Another Gibbs Sampler_, is a program for analyzing Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. [rjags](https://cran.r-project.org/web/packages/rjags/index.html) is R's interface to the `JAGS` library.
-    `JAGS` is required in some simulations `abn` can perform.
+  `GSL` is available throught the Advanced Package Tool, `APT`, simply type:
   
-    ### Installation
-
-    In prinicple, we could only install `JAGS` and have `rjags` be installed automatically as dependency when instlling `abn`, however, there are situations where `rjags` cannot find the `JAGS` libraries or `jags` at runtime, and therefore we recommend installing `rjags` directly after `JAGS` so to make sure that everying works as expected:
-  
-    - <details>
-        <summary>Ubuntu</summary>
-
-        `JAGS` is available throught the Advanced Package Tool, `APT`, simply type:
-
-        ```bash
-        sudo apt-get install jags
-        ```
-
-      </details>
-
-    - <details>
-        <summary>Fedora</summary>
-
-        `JAGS` must be installed from source.
-        Below are the steps to install `JAGS 4.3.2`.
-
-        ```bash
[…]
