+++ .github/workflows/onlabel_Macos_setup.yml
@@ -11,7 +11,7 @@ env:
   LABEL_SUCCESS: 'macosInstall::passed'
   LABEL_FAILURE: 'macosInstall::failed'
   BUILD_LOC: "./build"
-  BRANCH: ${{ github.ref_name }}
+  BRANCH: ${{ github.head_ref || github.ref_name }}
 
 permissions:
   packages: read
