+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -12,7 +12,7 @@ env:
   LABEL_FAILURE: 'fedoraInstall::failed'
   JAGS: '4.3.2'
   BUILD_LOC: "./build"
-  BRANCH: ${{ github.ref_name }}
+  BRANCH: ${{ github.head_ref || github.ref_name }}
 
 permissions:
   packages: read
