### 2024-07-14 4de2075
adding rjags install for ubunut
