+++ README.md
@@ -35,7 +35,7 @@ Officially supported is R version >= 4.4
 ## Prerequisites 
 
 <details>
-  <summary><b>Ubuntu</b></summary>
+  <summary><b><i>Ubuntu</i></b></summary>
  
 
   Most likely you have R installed already but if not run:
@@ -64,7 +64,7 @@ Officially supported is R version >= 4.4
 </details>
 
 <details>
-  <summary><b>Fedora</b></summary>
+  <summary><b><i>Fedora</i></b></summary>
 
   You will need C/C++ and Fortran compilers and various libraries:
  
@@ -101,7 +101,7 @@ Officially supported is R version >= 4.4
 </details>
 
 <details>
-  <summary><b>MacOS</b></summary>
+  <summary><b><i>MacOS</i></b></summary>
 
   The installation process on MacOS relies on [Homebrew](https://brew.sh/), head over to their site to see the installation process or simply open a terminal and run:
 
@@ -138,16 +138,16 @@ Some are C/C++ libraries and some are R packages that are not hosted on CRAN.
 
 Below is a list of these dependencies with details on how to install and configure them:
 
-<details>
-<summary><b>GSL</b></summary>
-
-[GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
-It is required to compile `abn`'s C/C++ code.
-
-#### Installation
-
 - <details>
-  <summary><b>Ubuntu</b></summary>
+  <summary><b>GSL</b></summary>
+  
+  [GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
+  It is required to compile `abn`'s C/C++ code.
+  
+  #### Installation
+  
+  <details>
+  <summary><b><i>Ubuntu</i></b></summary>
   
   `GSL` is available through the Advanced Package Tool, `APT`, simply type:
   
@@ -156,8 +156,8 @@ It is required to compile `abn`'s C/C++ code.
   ```
   
   </details>
-- <details>
-  <summary><b>Fedora</b></summary>
+  <details>
+  <summary><b><i>Fedora</i></b></summary>
   
   `GSL` is available through Fedoras standard package manager, `DNF`, simply type:
   
@@ -166,8 +166,8 @@ It is required to compile `abn`'s C/C++ code.
   ```
   
   </details>
-- <details>
-  <summary><b>MacOS</b></summary>
+  <details>
+  <summary><b><i>MacOS</i></b></summary>
   
   With Homebrew you can install the `GSL` binaries directly:
   
@@ -175,8 +175,8 @@ It is required to compile `abn`'s C/C++ code.
   brew install gsl
   ```
   </details>
-- <details>
-  <summary><b>Windows</b></summary>
+  <details>
+  <summary><b><i>Windows</i></b></summary>
   
   In Windows `GSL` is available a.o. through [cygwin](https://cygwin.com/index.html), which has a straight forward installation process.
   Either head over to the website, download and install the setup-x86_64.exe file, or use PowerShell:
@@ -188,38 +188,38 @@ It is required to compile `abn`'s C/C++ code.
   Start-Process -Wait -FilePath "C:\Program Files\cygwin\setup-x86_64.exe" -ArgumentList "/S" -PassThru
   ```
   </details>
-</details>
-<details><summary><b>JAGS & rjags</b></summary>
-
-[JAGS](https://mcmc-jags.sourceforge.io/), _Just Another Gibbs Sampler_, is a program for analyzing Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. [rjags](https://cran.r-project.org/web/packages/rjags/index.html) is R's interface to the `JAGS` library.
-`JAGS` is required in some simulations `abn` can perform.
-
-#### Installation
-
[…]
