### 2024-07-14 fbc544b
fixing some permissions for fedora
