### 2024-07-14 4bcf837
simply ignore the installation issue for html docs
