### 2024-07-14 616ce66
Fix: detailed installation instructions (#126)
* on label install check mac & windows

* fixing running condition

* installing dependencies for macos

* try installing gsl package

* dedicated window pipeline

* fixing os parameter

* dedicated macos workflow

* Revert "try installing gsl package"

This reverts commit b6678f53fe08036b47876b9b4dbfd3bbcd90e578.

* minor label formatting

* fixing container declaration

* installing jags

* adding jags path

* trying again with install

* installing rjags providing Jags home

* try with silent mode fore setup exe'

* adding more R versions

* in
[…]
