+++ README.md
@@ -16,62 +16,445 @@ It derives a directed acyclic graph (DAG) from empirical data that describes the
 The package provides routines for structure learning and parameter estimation of additive Bayesian network models.  
 
 # Installation
+[![Ubuntu Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Ubuntu_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Ubuntu_setup.yml)
+[![Fedora Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Fedora_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Fedora_setup.yml)
+[![MacOS Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Macos_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Macos_setup.yml)
+[![Windows Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Windows_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Windows_setup.yml)
 
-The most recent development version is available from [Github](https://github.com/furrer-lab/abn) and can be installed with:
+`abn` and its installation process relies on various software that might, or might not, be present in your system.
 
-```r
-devtools::install_github("furrer-lab/abn")
+The [Prerequisites](#prerequisites) section contains information about software that is needed but is (very) likely already installed in your system.
+In the [Dependencies](#dependencies) section you will find information about required software that is likely not already present.
+
+We recommend to go through both sections and follow the suggested installation process for your operating system before you try to install `abn`.
+
+## R version support
+
+Officially supported is R version >= 4.4
+
+## Prerequisites 
+
+<details>
+  <summary><b><i>Ubuntu</i></b></summary>
+ 
+
+  Most likely you have R installed already but if not run:
+ 
+  ```bash
+  apt-get install r-base
+  ```
+
+  To configure and build `abn` correctly we need `cmake` and `pkg-config`, as well as, the R `devtools` package:
+ 
+  ```bash
+  apt-get install pkg-config
+  apt-get install cmake
+  apt-get install r-cran-devtools
+  ```
+
+  Finally, we will use `wget` to download `JAGS` later, as well as, the development headers of `curl` and `openssl`:
+
+  ```bash
+  apt-get install wget
+  apt-get install libcurl4-openssl-dev
+  apt-get install libssl-dev
+
+  ```
+
+</details>
+
+<details>
+  <summary><b><i>Fedora</i></b></summary>
+
+  You will need C/C++ and Fortran compilers and various libraries:
+ 
+  ```bash
+  dnf install gcc
+  dnf install gcc-c++
+  dnf install gfortran
+  dnf install lapack*  # is needed on fedora server
+  dnf install blas*
+  dnf install atlas*
+  ```
+ 
+  Most likely you have R installed already but if not run:
+ 
+  ```bash
+  dnf install R
+  ```
+ 
+  To configure and build `abn` correctly we need `cmake` and `pkg-config`, as well as, the R `devtools` package:
+ 
+  ```bash
+  dnf install pkg-config
+  dnf install cmake
+  dnf install R-devtools
+  ```
+ 
+  Finally, we will use `wget` to download `JAGS` later, as well as, the development headers of `curl` and `openssl`:
+ 
+  ```bash
+  dnf install wget
+  dnf install libcurl-devel 
+  dnf install openssl-devel
+  ```
+</details>
+
+<details>
+  <summary><b><i>MacOS</i></b></summary>
+
[…]
