+++ DESCRIPTION
@@ -50,13 +50,14 @@ Imports:
     Rcpp,
     Rgraphviz,
     rjags,
-    stringi
+    stringi,
+    entropy,
+    moments
 Suggests: 
     bookdown,
     boot,
     brglm,
     devtools (>= 2.4.5),
-    entropy,
     ggplot2,
     gridExtra,
     INLA,
@@ -64,7 +65,6 @@ Suggests:
     Matrix (>= 1.6.3),
     MatrixModels (>= 0.5.3),
     microbenchmark,
-    moments,
     R.rsp,
     RhpcBLASctl,
     rmarkdown,

[…]
