### 2024-07-14 #6822 @j-i-l
Hi @abhishektiwari, hi @dhvalden ,

we've now updated the [Installation section](https://github.com/furrer-lab/abn?tab=readme-ov-file#installation) (see https://github.com/furrer-lab/abn/issues/129) and added (hopefully) exhaustive instructions to setup a working environment in various operating systems (Ubuntu, Fedora, MacOS and Windows).

We used various GitHub runners as starting point and implemented specific installation workflows to determine all the dependencies that need to be installed along with the precise commands to perform the insallation. Since GitHub does not provide a Fedora based runner, we used a container instead.

You will also find status badges in the [Installation section](https://github.com/furrer-lab/abn?tab=readme-ov-file#installation) that report the outcome of the installation procedure on these operating system and, on click, take you to the respective workflows:
[![Ubuntu Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Ubuntu_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Ubuntu_setup.yml) [![Fedora Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Fedora_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Fedora_setup.yml)[![MacOS Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Macos_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Macos_setup.yml)[![Windows Install](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Windows_setup.yml/badge.svg?branch=main)](https://github.com/furrer-lab/abn/actions/workflows/onlabel_Windows_setup.yml)

These workflows run on the latest versions of the operating systems (available as GitHub runners or Docker images) and run on `push` to `main`. They will allow us to spot quickly if the installation procedure needs updating.

Since we start from GitHub runners or a Docker image the Installation section might now contain some single steps that are not needed on the device of an end-user, however, trying to install already installed software should be less problematic than missing software.

Finally, the [recommended way of installing `abn`](https://github.com/furrer-lab/abn/blob/main/README.md#installing-from-github-recommended) encourages to install a specific tag (e.g. `3.1.1`) instead of targeting `main` directly. While this is generally recommendable, I would invite you to target `main` directly, as it already implements the revised Installation section, as well as, some fixes (e.g. #129).
