+++ README.md
@@ -23,48 +23,114 @@ The package provides routines for structure learning and parameter estimation of
 
 `abn` and its installation process relies on various software that might, or might not, be present in your system.
 
-The [Prerequisites](#prerequisites) section contains information about software that is needed but is (very) likely already installed in your system.
-In the [Dependencies](#dependencies) section you will find information about required software that is likely not already present.
+## Prior to installing
 
-We recommend to go through both sections and follow the suggested installation process for your operating system before you try to install `abn`.
+In order for `abn` to work correctly on your system some dependencies need to be installed first.
+The required steps to achieve this are specific to the operating system you are using and the following drop-down sections detail these steps for the most common ones: 
 
-## R version support
-
-Officially supported is R version >= 4.4
-
-## Prerequisites 
 
 <details>
-  <summary><b><i>Ubuntu</i></b></summary>
- 
+<summary><b><i>Ubuntu</i></b></summary>
 
-  Most likely you have R installed already but if not run:
- 
+#### Prerequisites 
+
+Most likely you have R installed already but if not run:
+
+```bash
+apt-get install r-base
+```
+
+To configure and build `abn` correctly we need `cmake` and `pkg-config`, as well as, the R `devtools` package:
+
+```bash
+apt-get install pkg-config
+apt-get install cmake
+apt-get install r-cran-devtools
+```
+
+Finally, we will use `wget` to download `JAGS` later, as well as, the development headers of `curl` and `openssl`:
+
+```bash
+apt-get install wget
+apt-get install libcurl4-openssl-dev
+apt-get install libssl-dev
+
+```
+
+#### Dependencies
+
+- **GSL**
+
+  [GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
+  It is required to compile `abn`'s C/C++ code.
+
+  `GSL` is available through the Advanced Package Tool, `APT`, simply type:
+  
   ```bash
-  apt-get install r-base
+  sudo apt-get install libgsl-dev
   ```
 
-  To configure and build `abn` correctly we need `cmake` and `pkg-config`, as well as, the R `devtools` package:
- 
+- **JAGS & rjags**
+  
+  [JAGS](https://mcmc-jags.sourceforge.io/), _Just Another Gibbs Sampler_, is a program for analyzing Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. [rjags](https://cran.r-project.org/web/packages/rjags/index.html) is R's interface to the `JAGS` library.
+  `JAGS` is required in some simulations `abn` can perform.
+
+  `JAGS` is available through the Advanced Package Tool, `APT`, simply type:
+  
   ```bash
-  apt-get install pkg-config
-  apt-get install cmake
-  apt-get install r-cran-devtools
+  sudo apt-get install jags
+  ```
+  And to install `rjags` open an R session and type:
+  
+  ```R
+  install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
+  library("rjags")
   ```
 
-  Finally, we will use `wget` to download `JAGS` later, as well as, the development headers of `curl` and `openssl`:
+- **INLA**
 
+  [INLA](https://www.r-inla.org/) is an R package that is not hosted on CRAN and thus needs to be installed separately.
+  `anb` uses `INLA` to fit some models. 
+  
[…]
