### 2024-07-16 cb69375
automatically regenerate configure file
