+++ .github/workflows/initiate_version_release.yml
@@ -88,6 +88,11 @@ jobs:
          sudo apt-get -y install autoconf
          autoreconf
 
+     - name: Update status badges in the readme file
+       run: | 
+         # updated the branch parameter for all badges
+         sed -i -e "s/?branch=release-${{ steps.release_version.outputs.previous_version}}/?branch=release-${{ steps.release_version.outputs.version}}/g" README.md
+
      - name: Create version release Pull Request
        id: release_pr
        # NOTE: We might want to perform this step with gh cli natively
@@ -97,6 +102,7 @@ jobs:
          commit-message: "Version ${{ steps.release_version.outputs.version }}"
          token: ${{ secrets.GITHUB_TOKEN }}
          add-paths: |
+           README.md
            NEWS.md
            DESCRIPTION
            configure.ac
