### 2024-07-17 furrer-lab/abn#123 @j-i-l
@dhvalden thanks for pointing this out.
It seems to me that on Fedora `JAGS` needs to be installed from source.

Also, when doing so, for some reason it gets installed under `/usr/local/lib` even on a 64-bit system which is a problem since libraries are expected to be in `/usr/local/lib64`.
Specifying the `libdir` explicitly for the configuration fixes this issue, so the installation procedure might look as follows:

```bash
wget -O /tmp/jags.tar.gz https://sourceforge.net/projects/mcmc-jags/files/JAGS/4.x/Source/JAGS-4.3.2.tar.gz/download
cd /tmp
tar -xf jags.tar.gz
cd /tmp/JAGS-4.3.2
./configure --libdir=/usr/local/lib64
make
sudo make install
```

With #126 we've added the required steps to install `JAGS` and the `rjags` for Fedora, and other common operating systems to the [README.md](https://github.com/furrer-lab/abn?tab=readme-ov-file#prior-to-installing).
