### 2024-07-18 408ffbc
adapted tests to expect a sparse matrix.
