### 2024-07-18 5b55e39
capture all version strings in release-x.y.z
