### 2024-07-18 236ec8a
127 run autoconf or autoreconf and commit configure file when creating new release branch (#130)
* automatically regenerate configure file

* avoid hardcoding branch

* Revert "avoid hardcoding branch"

This reverts commit 1e43957de682efc6c21a9c8fe92826a3a018d198.

* targeting the latest release branch for the status badge

* automatically update status badges

* capture all version strings in release-x.y.z
