### 2024-07-18 a3f22f5
attempt to rebuild vignettes on label. Addresses #50
