+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -1,0 +1,161 @@
+# Workflow triggered when we have a new release candidate
+# This action is adapted from https://github.com/t4d-gmbh/stubbed_versioning
+name: On Label rebuild vignettes
+
+on:
+  pull_request:
+    types: [ labeled ]
+
+env:
+  LABEL_CHECK: 'rebuildVigns::check'
+  LABEL_SUCCESS: 'rebuildVigns::passed'
+  LABEL_FAILURE: 'rebuildVigns::failed'
+  DOC_LOC: "./docs"
+
+jobs:
+  set_target:
+    if: ${{ github.event.label.name == 'rebuildVigns::check' }}
+    runs-on: ubuntu-latest
+    outputs:
+      label: ${{ steps.set_label.outputs.label }}
+    steps:
+      - id: set_label
+        run: |
+          echo "label=${{ env.LABEL_CHECK  }}" >> "$GITHUB_OUTPUT"
+
+  check_label_exist:
+    needs:
+      - set_target
+    runs-on: ubuntu-latest
+    env:
+      GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      OWNER: ${{ github.repository_owner }}
+      REPO: ${{ github.event.repository.name }}
+    permissions:
+      pull-requests: write
+      contents: write
+    steps:
+      - name: Assert labels ${{ env.LABEL_CHECK }} is defined
+        run: |
+          gh label create ${{ env.LABEL_CHECK }} --repo ${{ env.OWNER }}/${{ env.REPO }}
+        continue-on-error: true  # make sure the next steps run also on failure
+      - name: Assert labels ${{ env.LABEL_SUCCESS }} is defined
+        run: |
+          gh label create ${{ env.LABEL_SUCCESS }} --repo ${{ env.OWNER }}/${{ env.REPO }}
+        continue-on-error: true  # make sure the next steps run also on failure
+      - name: Assert labels ${{ env.LABEL_FAILURE }} is defined
+        run: |
+          gh label create ${{ env.LABEL_FAILURE }} --repo ${{ env.OWNER }}/${{ env.REPO }}
+        continue-on-error: true  # make sure the next steps run also on failure
+
+  vignette_build:
+    runs-on: ubuntu-latest
+    needs:
+      - set_target
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+    permissions:
+      contents: write
+      id-token: write
+    steps:
+      - uses: actions/checkout@v4
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          use-public-rspm: true
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: any::knitr, local::.
+      - name: Rebuild Vignettes
+        run: |
+          source("vignettes/precompile.R")
+        shell: Rscript {0}
+      - uses: actions/upload-artifact@v4
+        with:
+          path: vignettes/
+      - name: Commit compiled vignettes to branch
+        run: |
+          git config --global user.name 'GitHub Vignette Bot'
+          git config --global user.email 'vignettebot@github.com'
+          git commit -am "Automated Vignette Compilation"
+          git push
+
+  report_vignette_build:
+    if: ${{ (success() || failure()) }}
+    needs:
+      - vignette_build
+      - check_label_exist
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      pull-requests: write
+      repository-projects: write
+    env:
+      GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
[…]
