+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -52,6 +52,19 @@ jobs:
     runs-on: ubuntu-latest
     needs:
       - set_target
+    container:
+      image: ${{ vars.CONTAINER_SOURCE }}/${{ matrix.os }}/${{ matrix.compiler }}/${{ matrix.r-version }}/abn:${{ vars.CONTAINER_VERSION || 'latest' }}
+      credentials:
+        username: ${{ github.actor }}
+        password: ${{ secrets.GITHUB_TOKEN }}
+    env:
+      R_KEEP_PKG_SOURCE: yes  # NOTE Jo: not sure why this is here?
+    strategy:
+      fail-fast: false
+      matrix:
+        r-version: ['release']
+        os: ['debian']
+        compiler: ['gcc']
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     permissions:
@@ -59,12 +72,17 @@ jobs:
       id-token: write
     steps:
       - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          use-public-rspm: true
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          extra-packages: any::knitr, local::.
+      - name: Disable renv
+        run: |
+            renv::deactivate()
+        shell: Rscript {0}
+      - name: Configure ABN for installation
+        run: |
+          autoreconf
+      - name: Install package dependencies
+        run: |
+            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never')
+        shell: Rscript {0}
       - name: Rebuild Vignettes
         run: |
           source("vignettes/precompile.R")
