+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -93,8 +93,9 @@ jobs:
         run: |
           git config --global user.name 'GitHub Vignette Bot'
           git config --global user.email 'vignettebot@github.com'
+          currentBranch=$(git branch --show-current)
           git commit -am "Automated Vignette Compilation"
-          git push
+          git push origin HEAD:$currentBranch
 
   report_vignette_build:
     if: ${{ (success() || failure()) }}
