+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -93,11 +93,13 @@ jobs:
         with:
           path: vignettes/
       - name: Commit compiled vignettes to branch
-        run: |
-          git config --global user.name 'abn-cicd-helper[bot]' # retrieved from: https://api.github.com/users/abn-cicd-helper[bot]
-          git config --global user.email '170275362+abn-cicd-helper[bot]@users.noreply.github.com'
-          git commit -am "Automated Vignette Compilation"
-          git push origin HEAD:${{ env.BRANCH }}
+        uses: actions-js/push@master
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          author_email: '170275362+abn-cicd-helper[bot]@users.noreply.github.com'
+          author_name: 'abn-cicd-helper[bot]'
+          branch: ${{ env.BRANCH }}
+          message: 'Automated Vignette Compilation'
 
   report_vignette_build:
     if: ${{ (success() || failure()) }}
