+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -92,8 +92,7 @@ jobs:
       - uses: actions/upload-artifact@v4
         with:
           path: vignettes/
-      - name: Commit compiled vignettes to branch
-        uses: actions-js/push@master
+      - uses: actions-js/push@master
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           author_email: '170275362+abn-cicd-helper[bot]@users.noreply.github.com'
