+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -96,6 +96,7 @@ jobs:
         run: |
           git config --global user.name 'GitHub Vignette Bot'
           git config --global user.email 'vignettebot@github.com'
+          git config --global --add safe.directory /__w/abn/abn
           git commit -am "Automated Vignette Compilation"
           git push origin HEAD:${{ env.BRANCH }}
 
