### 2024-07-25 ef803d8
in container gh must be installed. On standart runners it would be by default.
