+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -77,6 +77,7 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
       - name: Checkout branch of PR
         run: |
+          apt-get update && apt-get install -y gh
           gh pr checkout ${{ env.EVENT }}
           git status
       - name: Disable renv
