### 2024-07-25 103fcbe
add specific sysreqs for INLA installation
