### 2024-07-25 ded5b4d
add specific sysreqs for INLA installation
