+++ DESCRIPTION
@@ -80,4 +80,4 @@ Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
-SystemRequirements: Gnu Scientific Library version >= 1.12, jpeg, gdal, geos, proj
+SystemRequirements: Gnu Scientific Library version >= 1.12, jpeg, gdal, geos, proj, udunits-2
