### 2024-07-25 b16b62d
Automated Vignette Compilation
