### 2024-07-25 6a16843
check the package to verify if INLA is installed
