+++ .github/workflows/onlabel_Ubuntu_setup_sysreqs.yml
@@ -1,99 +1,0 @@
-# Workflow is triggered on label and check installation on various operating systems
-# This action is adapted from https://github.com/t4d-gmbh/stubbed_versioning
-name: Ubuntu Install Sysreqs
-
-on:
-  push:
-    branches:
-      - "main"
-
-env:
-  JAGS: '4.3.2'
-  BUILD_LOC: "./build"
-  BRANCH: ${{ github.head_ref || github.ref_name }}
-
-permissions:
-  packages: read
-  contents: write
-  pull-requests: write
-  repository-projects: write
-
-jobs:
-  setup-and-install:
-    runs-on: ubuntu-latest
-    name: ubuntu-latest R release
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_KEEP_PKG_SOURCE: yes
-    steps:
-
-      - name: install R
-        run: |
-          sudo apt-get -y install r-base
-
-      - name: install config and build dependencies
-        run: |
-          sudo apt-get -y install pkg-config
-          sudo apt-get -y install cmake
-          sudo apt-get -y install r-cran-devtools
-
-      - name: fixing folder perms for local lib
-        run: |
-          sudo chown root:adm -R /usr/lib/R
-          sudo chmod g+w -R /usr/lib/R
-          sudo chown root:adm -R /usr/local/lib/R
-          sudo chmod g+w -R /usr/local/lib/R
-
-      - name: update R packages
-        run: |
-           update.packages(ask = FALSE, checkBuilt = TRUE)
-        shell: Rscript {0}
-
-      - name: install GSL
-        run: |
-          sudo apt-get -y install libgsl-dev
-
-      - name: install JAGS
-        run: |
-          sudo apt-get -y install jags
-
-      - name: install rjags
-        run: |
-            install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
-            library("rjags")
-        shell: Rscript {0}
-
-      - name: installing dependencies from BiocManager
-        run: |
-          if (!requireNamespace("BiocManager", quietly = TRUE))
-              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
-          BiocManager::install("Rgraphviz")
-          BiocManager::install("graph")
-        shell: Rscript {0}
-
-      - name: install from CRAN
-        run: |
-          install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
-          library('abn')
-        shell: Rscript {0}
-
-      - name: install from GitHub
-        run: |
-          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
-          library('abn')
-        shell: Rscript {0}
-
-      - name: fetch the repository from GitHub
-        uses: actions/checkout@v4
-
-      - name: deactivate the renv
-        run: |
-            renv::deactivate()
-        shell: Rscript {0}
-
-      - name: install from source
-        run: |
-            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never', repos=c(CRAN="https://cran.r-project.org"))
-            devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz', vignettes = FALSE)
-            install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
-        shell: Rscript {0}
