### 2024-07-25 cd63763
commit only recompiled vignettes but nothing else
