### 2024-07-25 e06efe6
try without gsl install (old DESC file)
