### 2024-07-25 f4b2e52
Revert "only update devtools - for speed"
This reverts commit c05b5f9ca988c680b95338d29965251936d812e8.
