### 2024-07-25 fbbb873
Automated Vignette Compilation
