+++ vignettes/multiprocessing.Rmd
@@ -43,7 +43,6 @@ We will use the `microbenchmark` package to measure the time it takes to compute
 
 ``` r
 library(abn)
-#> Error in library(abn): there is no package called 'abn'
 library(microbenchmark)
 
 set.seed(123456)
@@ -53,7 +52,6 @@ set.seed(123456)
 ``` r
 # Prepare data and parameters
 df <- FCV[, -c(13)]
-#> Error in eval(expr, envir, enclos): object 'FCV' not found
 mydists <- list(FCV = "binomial",
                 FHV_1 = "binomial",
                 C_felis = "binomial",
@@ -127,14 +125,14 @@ res <- microbenchmark(mleSinglecore = buildScoreCache(data.df = df,
                                                                             ncores = ncores,
                                                                             cluster.type = "FORK")),
                times = 25)
-#> Error in buildScoreCache(data.df = df, data.dists = mydists, method = "bayes", : could not find function "buildScoreCache"
 ```
 
 
 ``` r
 boxplot(res)
-#> Error in eval(expr, envir, enclos): object 'res' not found
 ```
+
+![plot of chunk benchmarkBoxPlot](benchmarkBoxPlot-1.png)
 
 The boxplot illustrates the time distribution for computing the score cache using different methods.
 
