+++ vignettes/parameter_learning.Rmd
@@ -12,7 +12,6 @@ vignette: >
 
 ``` r
 library(abn)
-#> Error in library(abn): there is no package called 'abn'
 ```
 
 In this vignette, we will discuss the parameter learning in the context of additive Bayesian network modelling with the package `abn`.
@@ -37,7 +36,6 @@ In this case, we need to provide the data, its distributions and the DAG as argu
 ``` r
 # Load an illustrative subset of the example data
 mydat <- ex0.dag.data[,c("b1","b2","b3","g1","b4","p2","p4")]
-#> Error in eval(expr, envir, enclos): object 'ex0.dag.data' not found
 
 # Specify the distributions of the nodes
 mydists <- list(b1="binomial",
@@ -58,7 +56,6 @@ mydag <- matrix(data=c( 0,0,1,0,0,0,0, # b1<-b3
                         0,0,0,0,0,0,0 #
 ), byrow=TRUE,ncol=7)
 colnames(mydag) <- rownames(mydag) <- names(mydat)
-#> Error in eval(expr, envir, enclos): object 'mydat' not found
 ```
 
 Fit the model to calculate the log marginal likelihood goodness of fit:
@@ -69,10 +66,10 @@ myres.c <- fitAbn(dag = mydag,
                   data.df = mydat,
                   data.dists = mydists,
                   method = "bayes")
-#> Error in fitAbn(dag = mydag, data.df = mydat, data.dists = mydists, method = "bayes"): could not find function "fitAbn"
 plot(myres.c)
-#> Error in eval(expr, envir, enclos): object 'myres.c' not found
 ```
+
+![plot of chunk unnamed-chunk-3](unnamed-chunk-3-1.png)
 
 # Examine the parameter estimates
 
@@ -81,7 +78,37 @@ We can examine the parameter estimates with the `print()` method for the class `
 
 ``` r
 print(myres.c)
-#> Error in eval(expr, envir, enclos): object 'myres.c' not found
+#> The ABN model was fitted using a Bayesian approach. The estimated modes (the highest posterior density values of the parameters) are:
+#> 
+#> $b1
+#> b1|(Intercept)          b1|b3 
+#>          0.217         -0.164 
+#> 
+#> $b2
+#> b2|(Intercept)          b2|b1 
+#>         0.0286         0.0465 
+#> 
+#> $b3
+#> b3|(Intercept) 
+#>         0.0267 
+#> 
+#> $g1
+#> g1|(Intercept)          g1|b4   g1|precision 
+#>        -0.0494         0.0915         1.0054 
+#> 
+#> $b4
+#> b4|(Intercept) 
+#>           0.16 
+#> 
+#> $p2
+#> p2|(Intercept) 
+#>           2.34 
+#> 
+#> $p4
+#> p4|(Intercept) 
+#>           2.32 
+#> 
+#> Number of nodes in the network:  7
 ```
 
 or by looking at the `coef` slot of the `abnFit` object returned by `fitAbn()`:
@@ -100,7 +127,6 @@ myres.c2 <- fitAbn(dag = mydag,
                   data.dists = mydists,
                   method = "bayes",
                   compute.fixed=TRUE)
-#> Error in fitAbn(dag = mydag, data.df = mydat, data.dists = mydists, method = "bayes", : could not find function "fitAbn"
 ```
 
 The `marginals` slot of the `abnFit` object returned by `fitAbn()` contains a list of all marginal posterior densities.
@@ -112,26 +138,21 @@ library(ggplot2)
 library(gridExtra)
 
 df1 <- as.data.frame(myres.c2$marginals$b1[["b1|(Intercept)"]])
[…]
