### 2024-07-25 ec70913
installing from source + checks
