+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -97,7 +97,7 @@ jobs:
           R CMD INSTALL .
       - name: Rebuild Vignettes
         run: |
-          cd vignettes
+          setwd("./vignettes")
           source("precompile.R")
         shell: Rscript {0}
       - uses: actions/upload-artifact@v4
