### 2024-07-26 9738699
adding curl as sysdep
