### 2024-07-26 3b4f175
trying INLA install via sysreq on Ubuntu
