+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -63,20 +63,6 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
-      - name: install INLA dependencies
-        run: |
-          sudo apt-get install -y --no-install-recommends libudunits2-dev
-          sudo apt -y install libcurl4-openssl-dev  # rCurl for (for INLA)
-          sudo apt-get -y install libssl-dev   # s2 package
-          sudo apt-get -y install libproj-dev  # sf package
-          sudo apt-get -y install libgeos-dev  # terra package
-          sudo apt-get -y install libgdal-dev  # terra package
-
-      - name: install INLA
-        run: |
-            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
-        shell: Rscript {0}
-
       - name: installing dependencies from BiocManager
         run: |
           if (!requireNamespace("BiocManager", quietly = TRUE))
