### 2024-07-26 0c0d650
trying INLA install via sysreq Fedora
