+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -73,21 +73,6 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
-      - name: install INLA dependencies
-        run: |
-          dnf -y install udunits2-devel
-          dnf -y install libcurl-devel  # rCurls package (for INLA)
-          dnf -y install openssl-devel  # s2 package
-          dnf -y install libjpeg-devel  # jpeg package
-          dnf -y install proj-devel     # sf package
-          dnf -y install geos-devel     # sf package
-          dnf -y install gdal-devel     # terra package
-
-      - name: install INLA
-        run: |
-            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
-        shell: Rscript {0}
-
       - name: installing dependencies from BiocManager
         run: |
           if (!requireNamespace("BiocManager", quietly = TRUE))
