### 2024-07-26 f1ec6a9
trying INLA install via sysreq Windows
