### 2024-07-26 cd10a07
upgrading ubuntu floavors
