+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -26,7 +26,12 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
+
     steps:
+
+      - name: Update and upgrade
+        run: |
+          sudo apt-get update && sudo apt-get upgrade
 
       - name: install R
         run: |
