### 2024-07-26 78ce274
simpler dependency statemnt (incl. suggests)
