+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -97,7 +97,6 @@ jobs:
           R CMD INSTALL .
       - name: Rebuild Vignettes
         run: |
-          setwd("./vignettes")
           source("precompile.R")
         shell: Rscript {0}
       - uses: actions/upload-artifact@v4
@@ -111,6 +110,7 @@ jobs:
           ls -la 
           ls -la vignettes/
           git add vignettes/
+          git add *.png
           git status
           git commit -m "Automated Vignette Compilation"
           git status
