+++ .github/workflows/onlabel_rebuild_vignettes.yml
@@ -99,6 +99,9 @@ jobs:
         run: |
           source("vignettes/precompile.R")
         shell: Rscript {0}
+      - name: Copy Figures to Vignettes folder
+        run: |
+          cp ./*.png vignettes/
       - uses: actions/upload-artifact@v4
         with:
           path: vignettes/
@@ -107,15 +110,10 @@ jobs:
           git config --global user.name 'GitHub Vignette Bot'
           git config --global user.email 'vignettebot@github.com'
           git config --global --add safe.directory /__w/abn/abn
-          ls -la 
-          ls -la vignettes/
           git add vignettes/
           git add *.png
-          git status
           git commit -m "Automated Vignette Compilation"
-          git status
           git push 
-          git status
 
   report_vignette_build:
     if: ${{ (success() || failure()) }}
