### 2024-07-26 807e8a8
bump R6 to dependency so it is installed on cueck_built
