### 2024-07-26 5fc768e
only fail on error
