+++ .github/workflows/onlabel_Ubuntu_setup.yml
@@ -106,5 +106,5 @@ jobs:
 
       - name: Run checks
         run: |
-          devtools::check(vignettes = FALSE)
+          devtools::check(vignettes=FALSE, error_on="error")
         shell: Rscript {0}
