+++ .github/workflows/onlabel_Fedora_setup.yml
@@ -111,5 +111,20 @@ jobs:
 
       - name: Run checks
         run: |
-          devtools::check(vignettes=FALSE, error_on="error")
+            options(crayon.enabled = TRUE)
+            cat("LOGNAME=", Sys.info()[["user"]], "\n", sep = "", file = Sys.getenv("GITHUB_ENV"), append = TRUE)
+            if (Sys.getenv("_R_CHECK_FORCE_SUGGESTS_", "") == "") Sys.setenv("_R_CHECK_FORCE_SUGGESTS_" = "false")
+            if (Sys.getenv("_R_CHECK_CRAN_INCOMING_", "") == "") Sys.setenv("_R_CHECK_CRAN_INCOMING_" = "false")
+            cat("check-dir-path=", file.path(getwd(), ("check")), "\n", file = Sys.getenv("GITHUB_OUTPUT"), sep = "", append = TRUE)
+            check_results <- rcmdcheck::rcmdcheck(".", args = (c("--no-manual", "--as-cran")), build_args = ("--no-manual", "--no-build-vignettes"), error_on = ("error"), check_dir = ("check"))
         shell: Rscript {0}
+
+      - name: Show checks outcome
+        run: |
+          echo ::group::Show testthat output
+          find check -name 'testthat.Rout*' -exec cat '{}' \; || true
+          echo ::endgroup::
+        shell: bash --noprofile --norc -e -o pipefail {0}
+        env:
+          BUILD_LOC: ./build
+          R_KEEP_PKG_SOURCE: yes
