+++ DESCRIPTION
@@ -39,7 +39,6 @@ URL: https://r-bayesian-networks.org/, https://github.com/furrer-lab/abn
 BugReports: https://github.com/furrer-lab/abn/issues
 Depends: 
     R (>= 4.0.0),
-    R6
 Imports: 
     doParallel,
     foreach,
