+++ .github/workflows/onlabel_Macos_setup.yml
@@ -101,7 +101,7 @@ jobs:
             if (Sys.getenv("_R_CHECK_FORCE_SUGGESTS_", "") == "") Sys.setenv("_R_CHECK_FORCE_SUGGESTS_" = "false")
             if (Sys.getenv("_R_CHECK_CRAN_INCOMING_", "") == "") Sys.setenv("_R_CHECK_CRAN_INCOMING_" = "false")
             cat("check-dir-path=", file.path(getwd(), ("check")), "\n", file = Sys.getenv("GITHUB_OUTPUT"), sep = "", append = TRUE)
-            check_results <- rcmdcheck::rcmdcheck(".", args = (c("--no-manual", "--as-cran")), build_args = ("--no-manual", "--no-build-vignettes"), error_on = ("error"), check_dir = ("check"))
+            check_results <- rcmdcheck::rcmdcheck(".", args = (c("--no-manual")), build_args = (c("--no-manual", "--no-build-vignettes")), error_on = ("error"), check_dir = ("check"))
         shell: Rscript {0}
 
       - name: Show checks outcome
