+++ DESCRIPTION
@@ -68,7 +68,8 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0),
     entropy,
-    moments
+    moments,
+    sf
 LinkingTo: 
     Rcpp,
     RcppArmadillo
