+++ DESCRIPTION
@@ -38,7 +38,7 @@ License: GPL (>= 3)
 URL: https://r-bayesian-networks.org/, https://github.com/furrer-lab/abn
 BugReports: https://github.com/furrer-lab/abn/issues
 Depends: 
-    R (>= 4.0.0),
+    R (>= 4.0.0)
 Imports: 
     doParallel,
     foreach,
@@ -60,7 +60,7 @@ Suggests:
     gridExtra,
     INLA,
     knitr,
-    Matrix (>= 1.6.3),
+    Matrix (== 1.6.0),
     MatrixModels (>= 0.5.3),
     microbenchmark,
     R.rsp,
@@ -69,7 +69,7 @@ Suggests:
     testthat (>= 3.0.0),
     entropy,
     moments,
-    sf
+    R6
 LinkingTo: 
     Rcpp,
     RcppArmadillo
