### 2024-07-28 4279bc3
attempt to rebuild vignettes on label (#137)
* attempt to rebuild vignettes on label. Addresses #50

* renamed the label

* use abn containers as basis.

* fixed env was defined twice.

* push to current branch

* changed branch name to github branch name.

* using repo scope token

* changed git user to abn-cicd-helper bot

* try commit and push with GH action...

* try again with the push action

* Revert "changed git user to abn-cicd-helper bot"

This reverts commit a1396d71cc3e952dc744444e97f3b5ef25b55a41.

* added directory as safe to fix dubious ownership warning.

* removed push origin

[…]
