### 2024-07-28 19c1575
remove dev running condition
