### 2024-07-28 f1e23e9
install with pak (and thus sysdeps)
