### 2024-07-28 e7b14ee
install curl for pak
