### 2024-07-28 65dfa27
separate step for pak install
