+++ .github/workflows/Fedora_setup_pak.yml
@@ -50,6 +50,11 @@ jobs:
           dnf -y install libcurl-devel  # needef for pak
           dnf -y install R-devtools
 
+      - name: install pak
+        run: |
+          install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
+        shell: Rscript {0}
+
       - name: install GSL
         run: |
           dnf -y install gsl-devel
