### 2024-07-28 d97ee3d
including optional deps (i.e. also Suggests)
