+++ .github/workflows/Fedora_setup_pak.yml
@@ -88,13 +88,9 @@ jobs:
           BiocManager::install("graph")
         shell: Rscript {0}
 
-      - name: adding repo for INLA to pak
-        run: |
-          pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
-        shell: Rscript {0}
-
       - name: pak install from GitHub
         run: |
+          pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
           pak::pkg_install("furrer-lab/abn@${{ env.BRANCH }}", dependencies=TRUE)
           library('abn')
         shell: Rscript {0}
