+++ .github/workflows/Fedora_setup_pak.yml
@@ -45,19 +45,12 @@ jobs:
 
       - name: install config and build dependencies
         run: |
-          dnf -y install pkg-config
-          dnf -y install cmake
           dnf -y install libcurl-devel  # needef for pak
-          dnf -y install R-devtools
 
       - name: install pak
         run: |
           install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
-
-      - name: install GSL
-        run: |
-          dnf -y install gsl-devel
 
       - name: install JAGS
         run: |
