### 2024-07-28 e3b2dfa
ditch biocm specialy install
