+++ .github/workflows/Fedora_setup_pak.yml
@@ -73,14 +73,6 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
-      - name: installing dependencies from BiocManager
-        run: |
-          if (!requireNamespace("BiocManager", quietly = TRUE))
-              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
-          BiocManager::install("Rgraphviz")
-          BiocManager::install("graph")
-        shell: Rscript {0}
-
       - name: pak install from GitHub
         run: |
           pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
