### 2024-07-28 73addef
adding jags as sysdep (for ubuntu)
