### 2024-07-28 4277d36
setup on Ubuntu with pak
