### 2024-07-28 c7f00f0
also adding libcurl for pak install
