+++ DESCRIPTION
@@ -78,7 +78,7 @@ VignetteBuilder:
 Additional_repositories: https://inla.r-inla-download.org/R/stable/
 Config/testthat/edition: 3
 Encoding: UTF-8
-LazyData: true
+LazyData: TRUE
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
 SystemRequirements: pkg-config, cmake, gsl, jpeg, gdal, geos, proj, udunits-2, openssl, libcurl, jags
