### 2024-07-30 3963a5a
since tests run isolated - load package first
