### 2024-07-31 e3a7294
pak based setup on Fedora
