+++ .github/workflows/Fedora_setup_pak.yml
@@ -6,6 +6,7 @@ on:
   push:
     branches:
       - "main"
+      - "117-elaborate-if-inla-jags-and-gsl-can-be-installed-via-systemrequirements"
 
 env:
   JAGS: '4.3.2'
@@ -95,6 +96,6 @@ jobs:
       - name: run the tests
         run: |
             pak::pak('r-lib/testthat')
+            library("abn")
             testthat::test_local(path='tests', load_package='none')
         shell: Rscript {0}
-
