### 2024-07-31 d90b07f
adding Macos pak setup
