### 2024-07-31 1c38573
better output for testthat
