+++ .github/workflows/Ubuntu_setup_pak.yml
@@ -9,8 +9,6 @@ on:
       - "117-elaborate-if-inla-jags-and-gsl-can-be-installed-via-systemrequirements"
 
 env:
-  JAGS: '4.3.2'
-  BUILD_LOC: "./build"
   BRANCH: ${{ github.head_ref || github.ref_name }}
 
 permissions:
@@ -53,6 +51,12 @@ jobs:
           install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
 
+      # - name: install from CRAN with pak
+      #   run: |
+      #     pak::pkg_install('abn', dependencies=TRUE)
+      #     library('abn')
+      #   shell: Rscript {0}
+
       - name: pak install from GitHub
         run: |
           pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
@@ -79,3 +83,11 @@ jobs:
             library("abn")
             testthat::test_local(path='tests', load_package='none')
         shell: Rscript {0}
+
+      - name: Show checks outcome
+        if: always()
+        run: |
+          echo ::group::Show testthat output
+          find check -name 'testthat.Rout*' -exec cat '{}' \; || true
+          echo ::endgroup::
+        shell: bash --noprofile --norc -e -o pipefail {0}
