### 2024-07-31 6c9dc1d
Windows setup with pak
