### 2024-08-05 b1f3a1b
updated instructions for Linux based systems
