+++ README.md
@@ -27,105 +27,39 @@ The package provides routines for structure learning and parameter estimation of
 
 ## Prior to installing
 
-In order for `abn` to work correctly on your system some dependencies need to be installed first.
-The required steps to achieve this are specific to the operating system you are using and the following drop-down sections detail these steps for the most common ones: 
+In order for `abn` to work correctly on your system some dependencies need to be installed.
+If you are on a Linux based system (most of) these dependencies are installed automatically for you when following the [pak](https://pak.r-lib.org/)-based installation procedure described in the [Installation section](#installing-from-github-recommended).
 
+For MacOS and Windows based system some more preparatory steps are required.
+
+The following paragraphs provide detailed instructions for the most common operating systems on the steps that need to be carried out prior to installing `abn`.
 
 <details>
 <summary><b><i>Ubuntu</i></b></summary>
 
 #### Prerequisites 
 
-Most likely you have R installed already but if not run:
+You presumably have R installed already, if not, open a terminal and type:
 
 ```bash
 apt-get install r-base
 ```
 
-To configure and build `abn` correctly we need `cmake` and `pkg-config`, as well as, the R `devtools` package:
+_Note: You might need to prepend `sudo ` to this command._
 
-```bash
-apt-get install pkg-config
-apt-get install cmake
-apt-get install r-cran-devtools
-```
+All you need for the installation is to have the R-package [pak](https://pak.r-lib.org/) installed.
+`pak` is installed like any other R-package, however, it relies on `curl` being installed on your system, so we make sure it is there:
 
-Finally, we will use `wget` to download `JAGS` later, as well as, the development headers of `curl` and `openssl`:
-
-```bash
-apt-get install wget
-apt-get install libcurl4-openssl-dev
-apt-get install libssl-dev
-
-```
-
-#### Dependencies
-
-- **GSL**
-
-  [GSL](https://www.gnu.org/software/gsl/), the _GNU Scientific Library_, is a numerical library for C/C++.
-  It is required to compile `abn`'s C/C++ code.
-
-  `GSL` is available through the Advanced Package Tool, `APT`, simply type:
-  
   ```bash
-  sudo apt-get install libgsl-dev
+  apt-get install libcurl4-openssl-dev
   ```
+  Now to install `pak` we start an R session and write:
 
-- **JAGS & rjags**
-  
-  [JAGS](https://mcmc-jags.sourceforge.io/), _Just Another Gibbs Sampler_, is a program for analyzing Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. [rjags](https://cran.r-project.org/web/packages/rjags/index.html) is R's interface to the `JAGS` library.
-  `JAGS` is required in some simulations `abn` can perform.
-
-  `JAGS` is available through the Advanced Package Tool, `APT`, simply type:
-  
-  ```bash
-  sudo apt-get install jags
-  ```
-  And to install `rjags` open an R session and type:
-  
   ```R
-  install.packages("rjags", configure.args="--enable-rpath", repos=c(CRAN="https://cran.r-project.org"))
-  library("rjags")
+  install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
   ```
 
-- **INLA**
-
-  [INLA](https://www.r-inla.org/) is an R package that is not hosted on CRAN and thus needs to be installed separately.
-  `anb` uses `INLA` to fit some models. 
-  
-  `INLA` relies on various other R packages and C/C++ libraries.
[…]
