### 2024-08-05 6b93c03
updated instructions for Linux based systems
