### 2024-08-06 af9ca06
adapting instructions for Mac&Windows
