### 2024-08-06 84efb1e
adpted installation instructions for pak
