### 2024-08-07 dbe701d
keep only intallation steps
