+++ .github/workflows/Windows_setup_pak.yml
@@ -88,11 +88,3 @@ jobs:
             library("abn")
             testthat::test_local(path='tests', load_package='none')
         shell: Rscript {0}
-
-      - name: Show checks outcome
-        if: always()
-        run: |
-          echo ::group::Show testthat output
-          find check -name 'testthat.Rout*' -exec cat '{}' \; || true
-          echo ::endgroup::
-        shell: bash --noprofile --norc -e -o pipefail {0}
