### 2024-08-07 aca0268
adapting setup batches
