+++ .github/workflows/onlabel_Windows_setup.yml
@@ -1,111 +1,0 @@
-# Workflow is triggered on label and check installation on various operating systems
-# This action is adapted from https://github.com/t4d-gmbh/stubbed_versioning
-name: Windows Install
-
-on:
-  push:
-    branches:
-      - "main"
-
-env:
-  BUILD_LOC: "./build"
-  BRANCH: ${{ github.head_ref || github.ref_name }}
-  JAGS_MAJORV: '4'
-  JAGS_VERSION: '4.3.1'
-  JAGS_URL: "https://sourceforge.net/projects/mcmc-jags/files/JAGS"
-
-permissions:
-  packages: read
-  contents: write
-  pull-requests: write
-  repository-projects: write
-
-jobs:
-  setup-and-install:
-    name: windows-latest R release
-    runs-on: windows-latest
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_KEEP_PKG_SOURCE: yes
-    steps:
-
-      - name: prepare windows
-        run: |
-            install.packages("devtools", repos=c(CRAN="https://cran.r-project.org"))
-        shell: Rscript {0}
-
-      - name: install GSL
-        run: |
-          Import-Module bitstransfer
-          New-Item -ItemType Directory -Force -Path "C:\Program Files\cygwin"
-          start-bitstransfer -source https://cygwin.com/setup-x86_64.exe "C:\Program Files\cygwin\setup-x86_64.exe"
-          Start-Process -Wait -FilePath "C:\Program Files\cygwin\setup-x86_64.exe" -ArgumentList "--quiet-mode --packages=gsl --site=https://cygwin.mirror.constant.com" -PassThru
-
-      - name: install JAGS
-        run: |
-          Import-Module bitstransfer
-          New-Item -ItemType Directory -Force -Path "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}"
-          start-bitstransfer -source ${{ env.JAGS_URL }}/${{ env.JAGS_MAJORV }}.x/Windows/JAGS-${{ env.JAGS_VERSION }}.exe/download  "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}\JAGS-${{ env.JAGS_VERSION }}.exe"
-          Start-Process -Wait -FilePath "C:\Program Files\JAGS\JAGS-${{ env.JAGS_VERSION }}\JAGS-${{ env.JAGS_VERSION }}.exe" -ArgumentList "/S" -PassThru
-
-      - name: install rjags separately
-        run: |
-            Sys.setenv(JAGS_HOME="C:/Program Files/JAGS/JAGS-${{ env.JAGS_VERSION }}")
-            install.packages("rjags", repos=c(CRAN="https://cran.r-project.org"))
-            library("rjags")
-        shell: Rscript {0}
-
-      - name: installing dependencies from BiocManager
-        run: |
-          if (!requireNamespace("BiocManager", quietly = TRUE))
-              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
-          BiocManager::install("Rgraphviz")
-          BiocManager::install("graph")
-        shell: Rscript {0}
-
-      - name: install from CRAN
-        run: |
-          install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
-          library('abn')
-        shell: Rscript {0}
-
-      - name: install from GitHub
-        run: |
-          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=TRUE, local=FALSE)
-          library('abn')
-        shell: Rscript {0}
-
-      - name: fetch the repository from GitHub
[…]
