### 2024-08-07 bd6ffd8
completing from source install
