### 2024-08-07 49f0a6d
add specific sysreqs for INLA installation (#139)
* add specific sysreqs for INLA installation

* run on this dev branch

* add udunits2 for fedora

* check the package to verify if INLA is installed

* we do not need to test the CRAN here

* run checks on HEAD of the branch

* no need to install from source

* try without gsl install (old DESC file)

* only update devtools - for speed

* Revert "only update devtools - for speed"

This reverts commit c05b5f9ca988c680b95338d29965251936d812e8.

* update only needed packages

* go back to updating all

* installing from source + checks

* adding 
[…]
