+++ .github/workflows/Fedora_setup.yml
@@ -44,13 +44,12 @@ jobs:
 
       - name: install config and build dependencies
         run: |
-          dnf -y install pkg-config
-          dnf -y install cmake
-          dnf -y install R-devtools
+          dnf -y install libcurl-devel  # needed for pak
 
-      - name: install GSL
+      - name: install pak
         run: |
-          dnf -y install gsl-devel
+          install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
+        shell: Rscript {0}
 
       - name: install JAGS
         run: |
@@ -73,38 +72,16 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
-      - name: install INLA dependencies
-        run: |
-          dnf -y install udunits2-devel
-          dnf -y install libcurl-devel  # rCurls package (for INLA)
-          dnf -y install openssl-devel  # s2 package
-          dnf -y install libjpeg-devel  # jpeg package
-          dnf -y install proj-devel     # sf package
-          dnf -y install geos-devel     # sf package
-          dnf -y install gdal-devel     # terra package
+      # - name: install from CRAN with pak
+      #   run: |
+      #     pak::pkg_install('abn', dependencies=TRUE)
+      #     library('abn')
+      #   shell: Rscript {0}
 
-      - name: install INLA
+      - name: pak install from GitHub
         run: |
-            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
-        shell: Rscript {0}
-
-      - name: installing dependencies from BiocManager
-        run: |
-          if (!requireNamespace("BiocManager", quietly = TRUE))
-              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
-          BiocManager::install("Rgraphviz")
-          BiocManager::install("graph")
-        shell: Rscript {0}
-
-      - name: install from CRAN
-        run: |
-          install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
-          library('abn')
-        shell: Rscript {0}
-
-      - name: install from GitHub
-        run: |
-          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
+          pak::pkg_install("furrer-lab/abn@${{ env.BRANCH }}", dependencies=TRUE)
           library('abn')
         shell: Rscript {0}
 
@@ -118,7 +95,5 @@ jobs:
 
       - name: install from source
         run: |
-            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never', repos=c(CRAN="https://cran.r-project.org"))
-            devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz', vignettes = FALSE)
-            install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
+            pak::local_install(dependencies=TRUE)
         shell: Rscript {0}
