+++ .github/workflows/Windows_setup.yml
@@ -29,9 +29,9 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
     steps:
 
-      - name: prepare windows
+      - name: install pak
         run: |
-            install.packages("devtools", repos=c(CRAN="https://cran.r-project.org"))
+          install.packages('pak', repos=c(CRAN="https://cran.r-project.org"))
         shell: Rscript {0}
 
       - name: install GSL
@@ -55,28 +55,16 @@ jobs:
             library("rjags")
         shell: Rscript {0}
 
-      - name: install INLA
-        run: |
-            install.packages("INLA", repos = c(CRAN = "https://cran.r-project.org", INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
-        shell: Rscript {0}
+      # - name: install from CRAN with pak
+      #   run: |
+      #     pak::pkg_install('abn', dependencies=TRUE)
+      #     library('abn')
+      #   shell: Rscript {0}
 
-      - name: installing dependencies from BiocManager
+      - name: pak install from GitHub
         run: |
-          if (!requireNamespace("BiocManager", quietly = TRUE))
-              install.packages("BiocManager", repos=c(CRAN="https://cran.r-project.org"))
-          BiocManager::install("Rgraphviz")
-          BiocManager::install("graph")
-        shell: Rscript {0}
-
-      - name: install from CRAN
-        run: |
-          install.packages('abn', repos=c(CRAN="https://cran.r-project.org"))
-          library('abn')
-        shell: Rscript {0}
-
-      - name: install from GitHub
-        run: |
-          devtools::install_github("furrer-lab/abn", ref="${{ env.BRANCH }}", dependencies=c("Depends", "Imports", "LinkingTo"), local=FALSE)
+          pak::repo_add(INLA = "https://inla.r-inla-download.org/R/stable/")
+          pak::pkg_install("furrer-lab/abn@${{ env.BRANCH }}", dependencies=TRUE)
           library('abn')
         shell: Rscript {0}
 
@@ -90,7 +78,12 @@ jobs:
 
       - name: install from source
         run: |
-            devtools::install_deps(pkg = '.', dependencies = TRUE, upgrade='never', repos=c(CRAN="https://cran.r-project.org"))
-            devtools::build(pkg = '.', path = '${{ env.BUILD_LOC }}/abn.tar.gz', vignettes = FALSE)
-            install.packages('${{ env.BUILD_LOC }}/abn.tar.gz', repos=NULL, type="source")
+            pak::local_install(dependencies=TRUE)
+        shell: Rscript {0}
+
+      - name: run the tests
+        run: |
+            pak::pak('r-lib/testthat')
+            library("abn")
+            testthat::test_local(path='tests', load_package='none')
         shell: Rscript {0}

[…]
