### 2024-08-07 708f0bb
always set the outcome label (#146)
