### 2024-08-07 furrer-lab/abn#147 @j-i-l
Completing DOI's
https://github.com/openjournals/joss-reviews/issues/6822#issuecomment-2143193458 reported some missing DOIs
