+++ vignettes/paper.bib
@@ -19,6 +19,7 @@
 @article{kratzer_information-theoretic_2018,
 	title = {Information-Theoretic Scoring Rules to Learn Additive Bayesian Network Applied to Epidemiology},
 	url = {http://arxiv.org/abs/1808.01126},
+  doi = {10.48550/arXiv.1808.01126},
 	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package abn that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
 	journaltitle = {{arXiv}:1808.01126 [cs, stat]},
 	author = {Kratzer, Gilles and Furrer, Reinhard},
@@ -366,6 +367,7 @@ Library},
 
 @Article{inla2013,
   title = {Bayesian computing with {INLA}: {N}ew features.},
+  doi = {10.1016/j.csda.2013.04.014},
   author = {Thiago G. Martins and Daniel Simpson and Finn Lindgren and H{\aa}vard Rue},
   year = {2013},
   volume = {67},
@@ -399,6 +401,7 @@ Library},
 @Article{testthat2011,
   author = {Hadley Wickham},
   title = {testthat: Get Started with Testing},
+  doi = {10.32614/rj-2011-002},
   journal = {The R Journal},
   year = {2011},
   volume = {3},
