### 2024-08-07 5756182
Assert all workflow correctly install the package (#145)
* micro

* silence some linters
