+++ .github/workflows/onlabel_lintingPlus.yml
@@ -85,7 +85,7 @@ jobs:
         shell: Rscript {0}
 
       - name: Lint root directory
-        run: lintr::lint_dir()
+        run: lintr::lint_dir(linters=lintr::with_defaults(line_length_linter=NULL, infix_spaces_linter=NULL, trailing_whitespace_linter=NULL))
         shell: Rscript {0}
         env:
           LINTR_ERROR_ON_LINT: true
