+++ NEWS.md
@@ -4,24 +4,11 @@
 
 > 2024-08-07
 
-### Dev
-
-* running code coverage with all tests ([#119](https://github.com/furrer-lab/abn/issues/119))
-
 ### Fix
 
-* detailed installation instructions ([#126](https://github.com/furrer-lab/abn/issues/126))
-
-### Fix
-
-* fetching correct tag for previous version
-
-### Pull Requests
-
-* Merge pull request [#118](https://github.com/furrer-lab/abn/issues/118) from furrer-lab/109-setup-sensible-code-coverage-report
-* Merge pull request [#115](https://github.com/furrer-lab/abn/issues/115) from furrer-lab/114-provide-guidance-on-prerequisite-for-installing-abn
-* Merge pull request [#110](https://github.com/furrer-lab/abn/issues/110) from furrer-lab/release-3.1.1
-
+* Simplified installation process ([#143](https://github.com/furrer-lab/abn/issues/143), [#142](https://github.com/furrer-lab/abn/issues/142), [#141](https://github.com/furrer-lab/abn/issues/141), [#125](https://github.com/furrer-lab/abn/issues/125), [#124](https://github.com/furrer-lab/abn/issues/124), [#123](https://github.com/furrer-lab/abn/issues/123), [#117](https://github.com/furrer-lab/abn/issues/117), [#114](https://github.com/furrer-lab/abn/issues/114))
+* Minor fixes in JOSS paper ([#147](https://github.com/furrer-lab/abn/issues/147), [#129](https://github.com/furrer-lab/abn/issues/129))
+* Some fixes and improvements in automation pipelines ([#140](https://github.com/furrer-lab/abn/issues/140), [#138](https://github.com/furrer-lab/abn/issues/138), [#132](https://github.com/furrer-lab/abn/issues/132), [#127](https://github.com/furrer-lab/abn/issues/127), [#111](https://github.com/furrer-lab/abn/issues/111))
 
 # abn 3.1.1
 [diff](https://github.com/furrer-lab/abn/compare/3.1.0...3.1.1)
