### 2024-08-08 2317cd5
Release version 3.1.2 (#149)
* Version 3.1.2

* improved installation process
* minor edits in the github workflows
* minor edits in the JOSS paper

---------

Co-authored-by: Jonas I. Liechti <j-i-l@t4d.ch>
