### 2024-08-08 #6822 @j-i-l
:wave: @abhishektiwari, @dhvalden, @crvernon,

We've just released [`abn` version `3.1.2`](https://github.com/furrer-lab/abn/releases/tag/3.1.2) :rocket:

The minor version increase contains a different approach for the installation process: We now recommend using ['pak'](https://pak.r-lib.org/) which is capable to pick-up system dependencies from the [`DESCRIPTION` file](https://github.com/furrer-lab/abn/blob/2317cd5cdca642fba5f14fd59355d3aae4ee3441/DESCRIPTION#L84).

In short, this approach simplifies drastically the installation process on Linux systems.
You will find the updated (and now much shorter) installation instructions [in the READE](https://github.com/furrer-lab/abn?tab=readme-ov-file#installation).

Hopefully, this leads to a better experience when using `abn`.
