### 2024-08-22 17d10e4
remove old installation procedure from paper
