+++ .github/workflows/Macos_setup.yml
@@ -30,6 +30,12 @@ jobs:
           brew install R
           brew install pkg-config
           brew install automake  # needed to run autoconf
+      
+      - name: temporal workaround for ragg
+        run: |
+          brew install curl
+          curl -L https://raw.githubusercontent.com/Homebrew/homebrew-core/d28afc3bfdb5d24eab02157bfffcc9e17ed6666c/Formula/f/freetype.rb > freetype.rb
+          brew install freetype.rb
 
       - name: install GSL
         run: |
