### 2024-09-06 afc1499
adding info about current but with 'ragg'
