+++ .github/workflows/Ubuntu_setup.yml
@@ -6,6 +6,7 @@ on:
   push:
     branches:
       - "main"
+      - "150-fails-to-build-dependency-ragg-when-installing"
 
 env:
   BRANCH: ${{ github.head_ref || github.ref_name }}
