### 2024-09-06 4162bc9
temporal fix for dependency build `ragg` (#151)
* downgrade freetype indirect dependency for ragg

* run on dev branch - to revert

* with absolute path

* trying versioning

* trying versioning

* using temp-fix in ragg

* target commit not branch

* adding info about current but with 'ragg'

* be more specific

* correcting spelling

* adding temp. workarounds

* running on dev branch - to revert
