### 2024-09-10 #6822 @dhvalden
@crvernon I'm sos sorry for the delay. I have updates, so i think i can close my review today. I must say that the documentation was greatly improved, thanks a lot to the authors of the package for that. I do have some comments (not issues, just comments) about the whole process tho. I spun up an Ubuntu LTS VM just to test the installation, instead of my fedora machine. The system dependencies are substantial! even in an Ubuntu system, which is very well populated from the start. I did manage to hunt down all the system dependencies, thanks if part of `pak` (great addition btw), but i must say that is definitely not very user friendly to install, and users less familiar with system dependencies might struggle.
Additionally, once i manage to find all the system dependencies, i noticed that abm itselft has a lot of dependencies (148 i think) and i don't really understand the reason for some of them. `viridis` for example, or some tidyverse packages. It seems that many of them are not direct dependencies.
Finally, i can report that the installation was succefull and that this time I was able to reproduce all the examples. I will open the checlist now and complete it. Again aim sos sorry for the delay. I was looking for ways to run the test in a way that didn;t change my system so much, i finally settle for VMs to be the best way.
