+++ README.md
@@ -298,16 +298,6 @@ Officially supported is R version >= 4.4
 
 ## Installing from GitHub (recommended)
 
-> [!WARNING]
-> Due to https://github.com/r-lib/ragg/issues/170 the R package [`ragg`](https://github.com/r-lib/ragg) does not build correctly.
-> Since `abn` depends on `ragg` the installation of `abn` might fail.
-> Therefore, we recommend installing a specific version of `ragg` as a temporal fix:
-> ```R
-> pak::pak("r-lib/ragg@0bc4e8c")
-> ```
-> Simply run this command before installing `abn` (see below) and you should be good to go.
-
-
 From GitHub you can install any version and/or state of the `abn` repository you want.
 We recommend to not directly install `main`, but to a specific version.
 Head over to our [version list](https://github.com/furrer-lab/abn/releases) to see which one is the latest version.
