+++ .github/workflows/Fedora_setup.yml
@@ -6,7 +6,6 @@ on:
   push:
     branches:
       - "main"
-      - "150-fails-to-build-dependency-ragg-when-installing"
 
 env:
   JAGS: '4.3.2'
@@ -78,11 +77,6 @@ jobs:
       #     pak::pkg_install('abn', dependencies=TRUE)
       #     library('abn')
       #   shell: Rscript {0}
-
-      - name: temporal workaround for ragg
-        run: |
-          pak::pak("r-lib/ragg@0bc4e8c")
-        shell: Rscript {0}
 
       - name: pak install from GitHub
         run: |
