+++ .github/workflows/Macos_setup.yml
@@ -6,7 +6,6 @@ on:
   push:
     branches:
       - "main"
-      - "150-fails-to-build-dependency-ragg-when-installing"
 
 env:
   BUILD_LOC: "./build"
@@ -69,11 +68,6 @@ jobs:
       #     pak::pkg_install('abn', dependencies=TRUE)
       #     library('abn')
       #   shell: Rscript {0}
-      
-      - name: temporal workaround for ragg
-        run: |
-          pak::pak("r-lib/ragg@0bc4e8c")
-        shell: Rscript {0}
 
       - name: pak install from GitHub
         run: |
