### 2024-09-11 #6822 @j-i-l
Thank you @dhvalden for the constructive feedback!

Out of curiosity: Did you install the CRAN version or the latest version from GitHub?
From `3.1.2` (currently `latest`) on installing it with `pak` on Ubuntu should take care of all system dependencies automatically.
_[Here](https://github.com/furrer-lab/abn/blob/411e45b9c30f5ea353e1b5debf7510e583da0b7c/.github/workflows/Ubuntu_setup.yml#L59-L64) is the relevant part in our Ubuntu test installation pipeline_.
Unfortunately `3.1.2` is not yet on CRAN and thus installing it from CRAN still requires the tedious process of installing the sys. deps. on your own.

Regarding your comment (and not issue :grimacing:  ) on the dependency list: Indeed, `abn` tracks quite a few indirect dependencies, which is not a problem _per se_, but there is, as you point out, not really a reason for keeping such an exhaustive dependency list.

We have a few things that we want to, and will, improve in the future and curating the dependency list is one of them.
However, since `pak` (or any other installation helper) takes care of installing the r-package dependencies, the length of this list should not affect the average user of `abn`.

Cycling back to the system dependencies: We had actually quite some discussions about that matter already beforehand and we decided to compile and maintain container images with completely setup environments.
They can be found here: https://github.com/furrer-lab/r-containers

We use these containers for testing, but they can also be used if someone wants to work with `abn` in a controlled, isolated and reproducible environment.
The container are updated monthly and we keep track of all installed packages and their upgrades (e.g. [dependency version changes August to September 2024](https://github.com/furrer-lab/r-containers/compare/2024.8.1...2024.9.1#diff-796553914f4d5b8d72edcfe70b9fed9cd278671fddfd8aca399b8c6712df925c)).

Since not every user is familiar with docker we decided to not actively advocate the usage of these containers for `abn`, but in light of your feedback maybe it would make sens to do so.
