### 2024-09-11 db36256
Revert "removing temp. fix for ragg" and adapt
This reverts commit 411e45b9c30f5ea353e1b5debf7510e583da0b7c.
