+++ .github/workflows/Macos_setup.yml
@@ -68,6 +68,11 @@ jobs:
       #     pak::pkg_install('abn', dependencies=TRUE)
       #     library('abn')
       #   shell: Rscript {0}
+      
+      - name: temporal workaround for ragg
+        run: |
+          pak::pak("r-lib/ragg")
+        shell: Rscript {0}
 
       - name: pak install from GitHub
         run: |
