+++ README.md
@@ -299,9 +299,9 @@ Officially supported is R version >= 4.4
 ## Installing from GitHub (recommended)
 
 > [!WARNING]
-> Until verion of `1.3.3` is released (see https://github.com/r-lib/ragg/issues/174) the R package [`ragg`](https://github.com/r-lib/ragg) does not build correctly.
+> Until the R package [`ragg`](https://github.com/r-lib/ragg) releases version `1.3.3` it might not build correctly.
 > Since `abn` depends on `ragg` the installation of `abn` might fail.
-> Therefore, we recommend installing the developer versin of `ragg` as a temporal fix:
+> Therefore, we recommend installing the developer version as a temporal fix:
 > ```R
 > pak::pak("r-lib/ragg")
 > ```
