### 2024-09-11 #6822 @dhvalden
dear @j-i-l!

Yes! I did used the recommended Github installation `pak` does highlight the missing system dependencies but does not install them automatically (at least in my case). Is still very useful tho. And not worries, my comments where just comments. Not really issues. I wanted to highlighted because in the future you might wanted to make the library more lightweight and this could be a path to it, as you already said. But again, with the help of `pak`, the installation process is much more friendly than before. I also assume that user wanting to use bayesian network packages will be familiar with some system dependencies missing in vanilla systems XD.
And yes! the container idea is amazing! Specially since `abn` is adding system dependencies that some people might not want to add. Thanks a lot for that and you should definitely promote that! Some people will run `abn` in HPC to which they have only partial privileges, and docker/singularity/apptainer images would be the solution for them.
Finally, sorry again for the delay. There really no excuse for it. I really enjoyed testing the library, and the latest changes are really good :)
