### 2024-09-16 #6822 @crvernon
:wave: @j-i-l and @matteodelucchi - this is looking good and we I am almost ready to move to publish.  Please address the following:

- [ ] In your paper, some of the references do not preserve capitalization as they should.  For example, in LINE 126, the "r" and "c++"  should both be capitalized.  There are others as well throughout the references that need to be addressed.  You can maintain formatting in your bib file using curly brackets around the characters that you wish to format.

Once the above has been addressed, we want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list matches the paper exactly). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version and the version identifier here
