### 2024-09-16 furrer-lab/abn#152 @j-i-l
Paper submission next steps
Copied from https://github.com/openjournals/joss-reviews/issues/6822#issuecomment-2352975830 in https://github.com/openjournals/joss-reviews/issues/6822:

-  [x] #153
-  [ ] #154
-  [ ] #155
- [ ] #156
