### 2024-09-17 04eac99
fixed capitalization in references, addressing #153.
