+++ vignettes/paper.bib
@@ -101,7 +101,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @article{kratzer_additive_2023,
-	title = {Additive Bayesian Network Modeling with the R Package abn},
+	title = {Additive Bayesian Network Modeling with the {R} Package {abn}},
 	volume = {105},
 	rights = {Copyright (c) 2023 Gilles Kratzer, Fraser I. Lewis, Arianna Comin, Marta Pittavino, Reinhard Furrer},
 	issn = {1548-7660},
@@ -119,7 +119,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @article{kalisch_causal_2012,
-	title = {Causal Inference Using Graphical Models with the R Package pcalg},
+	title = {Causal Inference Using Graphical Models with the {R} Package {pcalg}},
 	volume = {47},
 	rights = {Copyright (c) 2010 Markus Kalisch, Martin Mächler, Diego Colombo, Marloes H. Maathuis, Peter Bühlmann},
 	issn = {1548-7660},
@@ -155,7 +155,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @article{franzin_bnstruct_2017,
-	title = {bnstruct: an R package for Bayesian Network structure learning in the presence of missing data},
+	title = {{bnstruct}: an {R} package for Bayesian Network structure learning in the presence of missing data},
 	volume = {33},
 	issn = {1367-4803},
 	url = {https://doi.org/10.1093/bioinformatics/btw807},
@@ -169,11 +169,11 @@ Although the results needs to be interpreted with caution due to a small data se
 	author = {Franzin, Alberto and Sambo, Francesco and Di Camillo, Barbara},
 	urldate = {2024-03-26},
 	date = {2017-04-15},
-	file = {Full Text PDF:/home/matteo/Zotero/storage/U7DDDXPX/Franzin et al. - 2017 - bnstruct an R package for Bayesian Network struct.pdf:application/pdf},
+	file = {Full Text PDF:/home/matteo/Zotero/storage/U7DDDXPX/Franzin et al. - 2017 - bnstruct an {R} package for Bayesian Network struct.pdf:application/pdf},
 }
 
 @article{hojsgaard_graphical_2012,
-	title = {Graphical Independence Networks with the {gRain} Package for R},
+	title = {Graphical Independence Networks with the {gRain} Package for {R}},
 	volume = {46},
 	rights = {Copyright (c) 2009 Søren  Højsgaard},
 	issn = {1548-7660},
@@ -246,7 +246,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @Manual{rcore2024,
-  title = {R: A Language and Environment for Statistical Computing},
+  title = {{R}: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
@@ -265,7 +265,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @Manual{rjags2022,
-  title = {rjags: Bayesian Graphical Models using MCMC},
+  title = {rjags: Bayesian Graphical Models using {MCMC}},
   author = {Martyn Plummer},
   year = {2022},
   note = {R package version 4-13},
@@ -273,7 +273,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @Book{nnet2002,
-  title = {Modern Applied Statistics with S},
+  title = {Modern Applied Statistics with {S}},
   author = {W. N. Venables and B. D. Ripley},
   publisher = {Springer},
   edition = {Fourth},
[…]
