### 2024-09-17 51a58e7
more capitalization in references, addressing #153.
