+++ vignettes/paper.bib
@@ -20,7 +20,7 @@
 	title = {Information-Theoretic Scoring Rules to Learn Additive Bayesian Network Applied to Epidemiology},
 	url = {http://arxiv.org/abs/1808.01126},
   doi = {10.48550/arXiv.1808.01126},
-	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package abn that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
+	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package {abn} that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
 	journaltitle = {{arXiv}:1808.01126 [cs, stat]},
 	author = {Kratzer, Gilles and Furrer, Reinhard},
 	urldate = {2021-08-20},
@@ -107,7 +107,7 @@ Although the results needs to be interpreted with caution due to a small data se
 	issn = {1548-7660},
 	url = {https://doi.org/10.18637/jss.v105.i08},
 	doi = {10.18637/jss.v105.i08},
-	abstract = {The R package abn is designed to fit additive Bayesian network models to observational datasets and contains routines to score Bayesian networks based on Bayesian or information theoretic formulations of generalized linear models. It is equipped with exact search and greedy search algorithms to select the best network, and supports continuous, discrete and count data in the same model and input of prior knowledge at a structural level. The Bayesian implementation supports random effects to control for one-layer clustering. In this paper, we give an overview of the methodology and illustrate the package's functionality using a veterinary dataset concerned with respiratory diseases in commercial swine production.},
+	abstract = {The R package {abn} is designed to fit additive Bayesian network models to observational datasets and contains routines to score Bayesian networks based on Bayesian or information theoretic formulations of generalized linear models. It is equipped with exact search and greedy search algorithms to select the best network, and supports continuous, discrete and count data in the same model and input of prior knowledge at a structural level. The Bayesian implementation supports random effects to control for one-layer clustering. In this paper, we give an overview of the methodology and illustrate the package's functionality using a veterinary dataset concerned with respiratory diseases in commercial swine production.},
 	pages = {1--41},
 	journaltitle = {Journal of Statistical Software},
 	author = {Kratzer, Gilles and Lewis, Fraser I. and Comin, Arianna and Pittavino, Marta and Furrer, Reinhard},
