### 2024-09-17 d6555b4
build joss paper also on this branch
