### 2024-09-18 fee9172
152 paper submission next steps (#157)
* fixed capitalization in references, addressing #153.

* more capitalization in references, addressing #153.

* build joss paper also on this branch
