+++ vignettes/paper.bib
@@ -20,7 +20,7 @@
 	title = {Information-Theoretic Scoring Rules to Learn Additive Bayesian Network Applied to Epidemiology},
 	url = {http://arxiv.org/abs/1808.01126},
   doi = {10.48550/arXiv.1808.01126},
-	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package abn that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
+	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package {abn} that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
 	journaltitle = {{arXiv}:1808.01126 [cs, stat]},
 	author = {Kratzer, Gilles and Furrer, Reinhard},
 	urldate = {2021-08-20},
@@ -101,13 +101,13 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @article{kratzer_additive_2023,
-	title = {Additive Bayesian Network Modeling with the R Package abn},
+	title = {Additive Bayesian Network Modeling with the {R} Package {abn}},
 	volume = {105},
 	rights = {Copyright (c) 2023 Gilles Kratzer, Fraser I. Lewis, Arianna Comin, Marta Pittavino, Reinhard Furrer},
 	issn = {1548-7660},
 	url = {https://doi.org/10.18637/jss.v105.i08},
 	doi = {10.18637/jss.v105.i08},
[…]
