### 2024-09-18 e18dec9
updated the NEWS
