+++ NEWS.md
@@ -4,6 +4,7 @@
 
 > 2024-09-18
 
+* JOSS paper final edits ([#152](https://github.com/furrer-lab/abn/issues/152))
 
 # abn 3.1.2
 [diff](https://github.com/furrer-lab/abn/compare/3.1.1...3.1.2)
