### 2024-09-19 7f3e909
Release version 3.1.3 (#158)
* Version 3.1.3

* layouting

* updated the NEWS

* Updated citations

---------

Co-authored-by: matteodelucchi <37136726+matteodelucchi@users.noreply.github.com>
Co-authored-by: Matteo Delucchi <mail@matteodelucchi.com>
