### 2024-09-19 #6822 @matteodelucchi
> 👋 @j-i-l and @matteodelucchi - this is looking good and we I am almost ready to move to publish. Please address the following:
>
> * [ ]  In your paper, some of the references do not preserve capitalization as they should.  For example, in LINE 126, the "r" and "c++"  should both be capitalized.  There are others as well throughout the references that need to be addressed.  You can maintain formatting in your bib file using curly brackets around the characters that you wish to format.
>
> Once the above has been addressed, we want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So here is what we have left to do:
>
> * [ ]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on GitHub.
> * [ ]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list matches the paper exactly). You may also add the authors' ORCID.
> * [ ]   Please respond with the DOI of the archived version and the version identifier here
>
> I can then move forward with accepting the submission.

@crvernon We have cleaned the repository and published a new release, which we have linked on Zenodo. Here are the details:
DOI: [10.5281/zenodo.13788885](https://doi.org/10.5281/zenodo.13788885)
Version identifier: [3.1.3](https://github.com/furrer-lab/abn/releases/tag/3.1.3)
Please let me know if you need more information or any other details.

I just wanted to express my sincere gratitude to @abhishektiwari and @dhvalden for the thorough review. Your feedback has significantly improved the package once again. Thank you so much!
