### 2024-09-30 52db73c
Automated Vignette Compilation
