+++ vignettes/data_simulation.Rmd
@@ -12,7 +12,7 @@ vignette: >
 
 ``` r
 library(abn)
-#> abn version 3.1.1 (2024-05-22) is loaded.
+#> abn version 3.1.3 (2024-09-18) is loaded.
 #> To cite the package 'abn' in publications call: citation('abn').
 #> 
 #> Attaching package: 'abn'
