### 2024-09-30 856c6d5
159 capitalization in paper references (#160)
* Fixed capitalization in references

* Fixed capitalization in bibtex format of abn package citation.

* Automated Vignette Compilation

---------

Co-authored-by: GitHub Vignette Bot <vignettebot@github.com>
