+++ vignettes/paper.bib
@@ -1,6 +1,6 @@
 
 @article{kratzer_bayesian_2020,
-	title = {Bayesian Network Modeling Applied to Feline Calicivirus Infection Among Cats in Switzerland},
+	title = {{B}ayesian Network Modeling Applied to Feline Calicivirus Infection Among Cats in {S}witzerland},
 	volume = {7},
 	issn = {2297-1769},
 	url = {https://www.frontiersin.org/articles/10.3389/fvets.2020.00073/full},
@@ -17,7 +17,7 @@
 }
 
 @article{kratzer_information-theoretic_2018,
-	title = {Information-Theoretic Scoring Rules to Learn Additive Bayesian Network Applied to Epidemiology},
+	title = {Information-Theoretic Scoring Rules to Learn Additive {B}ayesian Network Applied to Epidemiology},
 	url = {http://arxiv.org/abs/1808.01126},
   doi = {10.48550/arXiv.1808.01126},
 	abstract = {Bayesian network modelling is a well adapted approach to study messy and highly correlated datasets which are very common in, e.g., systems epidemiology. A popular approach to learn a Bayesian network from an observational datasets is to identify the maximum a posteriori network in a search-and-score approach. Many scores have been proposed both Bayesian or frequentist based. In an applied perspective, a suitable approach would allow multiple distributions for the data and is robust enough to run autonomously. A promising framework to compute scores are generalized linear models. Indeed, there exists fast algorithms for estimation and many tailored solutions to common epidemiological issues. The purpose of this paper is to present an R package {abn} that has an implementation of multiple frequentist scores and some realistic simulations that show its usability and performance. It includes features to deal efficiently with data separation and adjustment which are very common in systems epidemiology.},
@@ -32,7 +32,7 @@
 }
 
 @article{pittavino_comparison_2017,
-	title = {Comparison between generalized linear modelling and additive Bayesian network; identification of factors associated with the incidence of antibodies against Leptospira interrogans sv Pomona in meat workers in New Zealand},
+	title = {Comparison between generalized linear modelling and additive {B}ayesian network; identification of factors associated with the incidence of antibodies against {L}eptospira interrogans sv {P}omona in meat workers in {N}ew {Z}ealand},
 	volume = {173},
 	issn = {0001-706X},
 	url = {https://www.sciencedirect.com/science/article/pii/S0001706X16308828},
@@ -55,7 +55,7 @@ The odds of Pomona infection was highest at stunning and hide removal. {PPE} did
 }
 
 @article{hartnack_additive_2019,
-	title = {Additive Bayesian networks for antimicrobial resistance and potential risk factors in non-typhoidal Salmonella isolates from layer hens in Uganda},
+	title = {Additive {B}ayesian networks for antimicrobial resistance and potential risk factors in non-typhoidal {S}almonella isolates from layer hens in {U}ganda},
 	volume = {15},
 	issn = {1746-6148},
 	url = {https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6591809/},
@@ -83,7 +83,7 @@ Although the results needs to be interpreted with caution due to a small data se
 }
 
 @article{delucchi_bayesian_2022,
-	title = {Bayesian network analysis reveals the interplay of intracranial aneurysm rupture risk factors},
[…]
