### 2024-09-30 #6822 @matteodelucchi
> @matteodelucchi - I still see issues with capitalization in your references:
>
> * [ ]  LINE 119: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 130: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 134: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 135: The "s" in "salmonella" should be capitalized
> * [ ]  LINE 135: The "u" in "uganda" should be capitalized
> * [ ]  LINE 143: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 145: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 153: The "s" in "switzerland" should be capitalized
> * [ ]  LINE 160: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 161: The "l" in "leptospira" should be capitalized
> * [ ]  LINE 161: The "p" in "pomona" should be capitalized
> * [ ]  LINE 161: The "n" and "z" in "new zealand" should be capitalized
> * [ ]  LINE 163: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 164: The "g" in "gibbs" should be capitalized
> * [ ]  LINE 166: The "b" in "bayesian" should be capitalized
> * [ ]  LINE 168: The "b" in "bayesian" should be capitalized

@crvernon Sorry, this took me longer than it should have. With this closed [PR](https://github.com/furrer-lab/abn/pull/160), the capitalization in the references should be correct now.
