### 2024-10-01 3678933
updated citation with JOSS publication
