+++ CITATION.cff
@@ -1,0 +1,48 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Delucchi
+  given-names: Matteo
+  orcid: "https://orcid.org/0000-0002-9327-1496"
+- family-names: Liechti
+  given-names: Jonas I.
+  orcid: "https://orcid.org/0000-0003-3447-3060"
+- family-names: Spinner
+  given-names: Georg R.
+  orcid: "https://orcid.org/0000-0001-9640-8155"
+- family-names: Furrer
+  given-names: Reinhard
+  orcid: "https://orcid.org/0000-0002-6319-2332"
+contact:
+- family-names: Furrer
+  given-names: Reinhard
+  orcid: "https://orcid.org/0000-0002-6319-2332"
+doi: 10.5281/zenodo.13788885
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Delucchi
+    given-names: Matteo
+    orcid: "https://orcid.org/0000-0002-9327-1496"
+  - family-names: Liechti
+    given-names: Jonas I.
+    orcid: "https://orcid.org/0000-0003-3447-3060"
+  - family-names: Spinner
+    given-names: Georg R.
+    orcid: "https://orcid.org/0000-0001-9640-8155"
+  - family-names: Furrer
+    given-names: Reinhard
+    orcid: "https://orcid.org/0000-0002-6319-2332"
+  date-published: 2024-09-30
+  doi: 10.21105/joss.06822
+  issn: 2475-9066
+  issue: 101
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 6822
+  title: Additive Bayesian Networks
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.06822"
+  volume: 9
+title: Additive Bayesian Networks
